% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plug-and-play.R
\name{add_meas_BrS_case_NoNest_reg_discrete_predictor_Slice_jags}
\alias{add_meas_BrS_case_NoNest_reg_discrete_predictor_Slice_jags}
\title{add likelihood component for a BrS measurement slice among cases}
\usage{
add_meas_BrS_case_NoNest_reg_discrete_predictor_Slice_jags(s, Mobs, prior,
  cause_list, ppd = NULL)
}
\arguments{
\item{s}{the slice}

\item{Mobs}{See \code{data_nplcm} described in \code{\link{nplcm}}}

\item{prior}{Prior specifications.}

\item{cause_list}{the list of causes in \code{data_nplcm} described in \code{\link{nplcm}}}

\item{ppd}{Default is NULL; Set to TRUE for enabling posterior predictive checking.}
}
\value{
a list of two elements: the first is \code{plug}, the .bug code; 
the second is \code{parameters} that stores model parameters introduced by this 
plugged measurement slice
}
\description{
regression model with no nested subclasses; discrete predictors
}
\seealso{
Other likelihood specification functions: \code{\link{add_meas_BrS_case_Nest_Slice_jags}},
  \code{\link{add_meas_BrS_case_Nest_Slice}},
  \code{\link{add_meas_BrS_case_NoNest_Slice_jags}},
  \code{\link{add_meas_BrS_case_NoNest_Slice}},
  \code{\link{add_meas_BrS_case_NoNest_reg_Slice_jags}},
  \code{\link{add_meas_BrS_ctrl_Nest_Slice}},
  \code{\link{add_meas_BrS_ctrl_NoNest_Slice}},
  \code{\link{add_meas_BrS_ctrl_NoNest_reg_Slice_jags}},
  \code{\link{add_meas_BrS_ctrl_NoNest_reg_discrete_predictor_Slice_jags}},
  \code{\link{add_meas_BrS_param_Nest_Slice_jags}},
  \code{\link{add_meas_BrS_param_Nest_Slice}},
  \code{\link{add_meas_BrS_param_NoNest_Slice_jags}},
  \code{\link{add_meas_BrS_param_NoNest_Slice}},
  \code{\link{add_meas_BrS_param_NoNest_reg_Slice_jags}},
  \code{\link{add_meas_BrS_param_NoNest_reg_discrete_predictor_Slice_jags}},
  \code{\link{add_meas_BrS_subclass_Nest_Slice}},
  \code{\link{add_meas_SS_case}},
  \code{\link{add_meas_SS_param}}

Other plug-and-play functions: \code{\link{add_meas_BrS_case_Nest_Slice_jags}},
  \code{\link{add_meas_BrS_case_Nest_Slice}},
  \code{\link{add_meas_BrS_case_NoNest_Slice_jags}},
  \code{\link{add_meas_BrS_case_NoNest_Slice}},
  \code{\link{add_meas_BrS_case_NoNest_reg_Slice_jags}},
  \code{\link{add_meas_BrS_ctrl_Nest_Slice}},
  \code{\link{add_meas_BrS_ctrl_NoNest_Slice}},
  \code{\link{add_meas_BrS_ctrl_NoNest_reg_Slice_jags}},
  \code{\link{add_meas_BrS_ctrl_NoNest_reg_discrete_predictor_Slice_jags}},
  \code{\link{add_meas_BrS_param_Nest_Slice_jags}},
  \code{\link{add_meas_BrS_param_Nest_Slice}},
  \code{\link{add_meas_BrS_param_NoNest_Slice_jags}},
  \code{\link{add_meas_BrS_param_NoNest_Slice}},
  \code{\link{add_meas_BrS_param_NoNest_reg_Slice_jags}},
  \code{\link{add_meas_BrS_param_NoNest_reg_discrete_predictor_Slice_jags}},
  \code{\link{add_meas_BrS_subclass_Nest_Slice}},
  \code{\link{add_meas_SS_case}},
  \code{\link{add_meas_SS_param}}
}
\concept{likelihood specification functions}
\concept{plug-and-play functions}
