% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plug-and-play.R
\name{add_meas_BrS_ctrl_NoNest_Slice}
\alias{add_meas_BrS_ctrl_NoNest_Slice}
\title{add a likelihood component for a BrS measurement slice among controls (conditional independence)}
\usage{
add_meas_BrS_ctrl_NoNest_Slice(s, Mobs, cause_list)
}
\arguments{
\item{s}{the slice}

\item{Mobs}{See \code{data_nplcm} described in \code{\link{nplcm}}}

\item{cause_list}{the list of causes in \code{data_nplcm} described in \code{\link{nplcm}}}
}
\value{
a list of two elements: the first is \code{plug}, the .bug code; the second is \code{parameters}
that stores model parameters introduced by this plugged measurement slice
}
\description{
add a likelihood component for a BrS measurement slice among controls (conditional independence)
}

