% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plug-and-play.R
\name{add_meas_BrS_param_NoNest_Slice}
\alias{add_meas_BrS_param_NoNest_Slice}
\title{add parameters for a BrS measurement slice among cases and controls (conditional independence)}
\usage{
add_meas_BrS_param_NoNest_Slice(s, Mobs, cause_list)
}
\arguments{
\item{s}{the slice}

\item{Mobs}{See \code{data_nplcm} described in \code{\link{nplcm}}}

\item{cause_list}{the list of causes in \code{data_nplcm} described in \code{\link{nplcm}}}
}
\value{
a list of two elements: the first is \code{plug}, the .bug code; the second is \code{parameters}
that stores model parameters introduced by this plugged measurement slice
}
\description{
add parameters for a BrS measurement slice among cases and controls (conditional independence)
}
\seealso{
Other likelihood specification functions: \code{\link{add_meas_BrS_case_Nest_Slice_jags}},
  \code{\link{add_meas_BrS_case_Nest_Slice}},
  \code{\link{add_meas_BrS_case_NoNest_Slice_jags}},
  \code{\link{add_meas_BrS_case_NoNest_Slice}},
  \code{\link{add_meas_BrS_ctrl_Nest_Slice}},
  \code{\link{add_meas_BrS_ctrl_NoNest_Slice}},
  \code{\link{add_meas_BrS_param_Nest_Slice_jags}},
  \code{\link{add_meas_BrS_param_Nest_Slice}},
  \code{\link{add_meas_BrS_param_NoNest_Slice_jags}},
  \code{\link{add_meas_BrS_subclass_Nest_Slice}},
  \code{\link{add_meas_SS_case}},
  \code{\link{add_meas_SS_param}}

Other plug-and-play functions: \code{\link{add_meas_BrS_case_Nest_Slice_jags}},
  \code{\link{add_meas_BrS_case_Nest_Slice}},
  \code{\link{add_meas_BrS_case_NoNest_Slice_jags}},
  \code{\link{add_meas_BrS_case_NoNest_Slice}},
  \code{\link{add_meas_BrS_ctrl_Nest_Slice}},
  \code{\link{add_meas_BrS_ctrl_NoNest_Slice}},
  \code{\link{add_meas_BrS_param_Nest_Slice_jags}},
  \code{\link{add_meas_BrS_param_Nest_Slice}},
  \code{\link{add_meas_BrS_param_NoNest_Slice_jags}},
  \code{\link{add_meas_BrS_subclass_Nest_Slice}},
  \code{\link{add_meas_SS_case}},
  \code{\link{add_meas_SS_param}}
}

