% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plug-and-play.R
\name{add_meas_SS_param}
\alias{add_meas_SS_param}
\title{add parameters for a SS measurement slice among cases (conditional independence)}
\usage{
add_meas_SS_param(nslice, Mobs, prior, cause_list)
}
\arguments{
\item{nslice}{the total number of SS measurement slices}

\item{Mobs}{see \code{data_nplcm} described in \code{\link{nplcm}}}

\item{prior}{see \code{model_options} described in \code{\link{nplcm}}}

\item{cause_list}{the list of causes in \code{model_options} described in \code{\link{nplcm}}}
}
\value{
a list of two elements: the first is \code{plug}, the .bug code; the second is \code{parameters}
that stores model parameters introduced by this plugged measurement slice
}
\description{
add parameters for a SS measurement slice among cases (conditional independence)
}

