% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{as.matrix_or_vec}
\alias{as.matrix_or_vec}
\title{convert one column data frame to a vector}
\usage{
as.matrix_or_vec(x)
}
\arguments{
\item{x}{an one-column data.frame}
}
\value{
a vector
}
\description{
convert one column data frame to a vector
}
\details{
WinBUGS/JAGS cannot accept a dataframe with one column; This function
converts it to a vector, which WinBUGS/JAGS will allow.
}
