% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/assign_model.R
\name{assign_model}
\alias{assign_model}
\title{Choose a model to fit}
\usage{
assign_model(Mobs, Y, X, model_options, silent = TRUE)
}
\arguments{
\item{Mobs}{See \code{\link{nplcm}} function.}

\item{Y}{See \code{\link{nplcm}} function.}

\item{X}{See \code{\link{nplcm}} function.}

\item{model_options}{See \code{\link{nplcm}} function.}

\item{silent}{Default is \code{TRUE}: print assigned
model descriptions on screen; otherwise, \code{FALSE}.}
}
\value{
A list of model information:
\itemize{
\item \code{measurement}
\itemize{
\item \code{quality} e.g. "BrS+SS" indicates both BrS and SS measures are
available
\item \code{SSonly} \code{TRUE} for existence of pathogens with only SS measures;
otherwise, \code{FALSE};
\item \code{nest} \code{TRUE} for conditional dependent model; \code{FALSE}
for conditional independent model;
}
\item \code{reg}
\itemize{
\item \code{do_FPR_reg} \code{TRUE} for allowing FPR to be covariate-dependent; \code{FALSE} otherwise;
\item \code{do_Eti_reg} \code{TRUE} for allowing etiology to be covariate-dependent; \code{FALSE} otherwise;
}
}
}
\description{
\code{assign_model} recognizes the model to fit as requested by user.
\code{assign_model} will also inspect the actual data supplied to this function
 and check if the data conform to user's requested model. The following features of data and user
inputs are checked: 1) available measurement quality types, i.e., gold-, silver- or bronze-standard or
any combinations; 2) model for false positive rates: covariate-dependent or not;
3) model for etiology: covariate-dependent or not.
}

