% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/clean-perch-data.R
\name{assign_taxo}
\alias{assign_taxo}
\title{assign taxonomy information to every slice of measurements}
\usage{
assign_taxo(meas_object, dir_taxo)
}
\arguments{
\item{meas_object}{measurement object constructed by \link{make_meas_object}}

\item{dir_taxo}{file path to the .csv file storing two columns of information:
\code{pathogen}, and \code{pathogen_type}.}
}
\value{
a vector of factors of taxonomy information (currently for "B","F","V").
The names of the vector are pathogen names.
}
\description{
assign taxonomy information to every slice of measurements
}

