% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean-perch-data.R
\name{assign_taxo_cause_list}
\alias{assign_taxo_cause_list}
\title{assign taxonomy information to cause list}
\usage{
assign_taxo_cause_list(cause_list, dir_taxo = NULL)
}
\arguments{
\item{cause_list}{See \code{cause_list} in \code{model_options$likelihood} (see 
parameter settings in \code{\link{nplcm}})}

\item{dir_taxo}{File path to the .csv file storing two columns of information:
\code{pathogen}, and \code{pathogen_type}.}
}
\value{
A vector of factors of taxonomy information (currently for "B","F","V"). 
The names of the vector are pathogen names.
}
\description{
assign taxonomy information to cause list
}

