% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/baker.R
\docType{package}
\name{baker}
\alias{baker}
\alias{baker-package}
\title{baker: A package for Bayesian inference of nested partially-latent
class models (NPLCM).}
\description{
The baker package provides functionalities for cleaning PERCH data (the
motivating study), exploratory data analysis, fitting NPLCM, posterior
summaries and model checking.
}
\section{baker functions}{

The baker functions ...
}

