% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{baker_help}
\alias{baker_help}
\title{Helping messages for setting up analyses}
\usage{
baker_help()
}
\value{
Text messages on the Operating System (OS) and how to set up \code{mcmc_options}
to call the right Bayesian fitting functions.
}
\description{
Helping messages for setting up analyses
}

