% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{beta_plot}
\alias{beta_plot}
\title{Plot beta density}
\usage{
beta_plot(a, b)
}
\arguments{
\item{a}{The first parameter}

\item{b}{The second parameter}
}
\value{
None
}
\description{
Plot beta density
}

