% Generated by roxygen2 (4.0.2): do not edit by hand
\name{bin2dec}
\alias{bin2dec}
\title{Convert 0/1 binary coded sequence into decimal digits}
\usage{
bin2dec(binary_vector)
}
\arguments{
\item{binary_vector}{a binary number}
}
\value{
a decimal number
}
\description{
Useful when try to list all the binary patterns. One can group the binary
sequences according to their equivalent decimal values.
}

