% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_dir_create}
\alias{check_dir_create}
\title{check existence and create folder if non-existent}
\usage{
check_dir_create(path)
}
\arguments{
\item{path}{Folder path to check and create if not there.}
}
\value{
the same returned values for \code{\link[=dir.create]{dir.create()}}
}
\description{
check existence and create folder if non-existent
}
\examples{
\donttest{
run_example <- function(){
 xx <- tempdir()
 check_dir_create(xx)
 on.exit(unlink(xx, recursive = TRUE), add = TRUE) 
}
run_example()
}

}
