% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/check-pairwise-logOR-two-folders.R
\name{check_pairwise_logOR_two_folders}
\alias{check_pairwise_logOR_two_folders}
\title{Posterior predictive checking for nested partially latent class models -
pairwise log odds ratio (parallel comparisons for two folders)}
\usage{
check_pairwise_logOR_two_folders(DIR_NPLCM1, DIR_NPLCM2)
}
\arguments{
\item{DIR_NPLCM1}{File path to the first folder that stores results from npLCM fit.}

\item{DIR_NPLCM2}{File path to the second folder}
}
\value{
A figure of posterior predicted log odds ratio compared with the observed
log odds ratio for the BrS data. The function generates this figure in your working directory automatically.
}
\description{
At each MCMC iteration, we generate a new data set based on the model and
parameter values at that iteration. The sample size of the new data set equals
that of the actual data set, i.e. the same number of cases and controls.
}

