% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/assign-model.R
\name{check_SS_grp}
\alias{check_SS_grp}
\title{check if SS measurements have stratified TPRs}
\usage{
check_SS_grp(model_options)
}
\arguments{
\item{model_options}{See \link{nplcm}}
}
\value{
TRUE for stratified TPRs; FALSE otherwise
}
\description{
check if SS measurements have stratified TPRs
}

