% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean-perch-data.R
\name{clean_combine_subsites}
\alias{clean_combine_subsites}
\title{Combine subsites in raw PERCH data set}
\usage{
clean_combine_subsites(raw_meas_dir, subsites_list, newsites_vec)
}
\arguments{
\item{raw_meas_dir}{The file path to the raw data file (.csv)}

\item{subsites_list}{The list of subsite group names. Each group is a vector of
subsites to be combined}

\item{newsites_vec}{A vector of new site names. It has the same length as
\code{"subsites_list"}}
}
\value{
A data frame with combined sites
}
\description{
In the Actual PERCH data set, a study site may have multiple subsites.
\code{clean_combine_subsites} combines all the study subjects from the same site.
}
