% Generated by roxygen2 (4.0.2): do not edit by hand
\name{combined_visualization}
\alias{combined_visualization}
\title{combine results from two folder, specially made for data visualization:}
\usage{
combined_visualization(DIR_NPLCM, DIR_PLCM, npat = 16)
}
\arguments{
\item{DIR_NPLCM}{The folder path to results from npLCM fit}

\item{DIR_PLCM}{The folder path to results from pLCM fit}

\item{npat}{The number of bronze-standard patterns to display individual diagnosis
plots}
}
\value{
Individual plots comparing two nested models: pLCM and npLCM
}
\description{
This function creates visualizations that are convenient for comparing the
individual prediction from two models: pLCM and npLCM.
}

