% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_functionals_of_model.R
\name{compute_logOR_single_and_other_cause}
\alias{compute_logOR_single_and_other_cause}
\title{Calculate marginal log odds ratios (with other causes)}
\usage{
compute_logOR_single_and_other_cause(set_parameter)
}
\arguments{
\item{set_parameter}{True model parameters in the npLCM specification}
}
\value{
a matrix
}
\description{
Calculate marginal log odds ratios (with other causes)
}
