% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_functionals_of_model.R
\name{compute_marg_PR_nested_reg}
\alias{compute_marg_PR_nested_reg}
\title{compute positive rates for nested model with subclass mixing weights that are the same across
\code{Jcause} classes for each person (people may have different weights.)}
\usage{
compute_marg_PR_nested_reg(ThetaBS, PsiBS, pEti_mat, subwt_mat, case, template)
}
\arguments{
\item{ThetaBS}{True positive rates for JBrS measures (rows) among K subclasses (columns)}

\item{PsiBS}{False positive rates; dimension same as above}

\item{pEti_mat}{a matrix of etiology pies for N subjects (rows) and Jcause causes (columns)
rows sum to ones.}

\item{subwt_mat}{a matrix of subclass weights for cases and controls. N by K. Rows sum
to ones.}

\item{case}{a N-vector of 1s (cases) and 0s (controls)}

\item{template}{a binary matrix with Jcause+1 rows (Jcause classes of cases and 1 class of controls)
and JBrS columns for the Bronze-standard measurement (say, pick one type/slice).
The ones in each row indicate the measurements that will show up more frequently in cases given the cause.}
}
\value{
a matrix of values between 0 and 1 (need not to have row sums of ones); of dimension (number of subjects, dimension of the bronze-standard
measurement slice).
}
\description{
compute positive rates for nested model with subclass mixing weights that are the same across
\code{Jcause} classes for each person (people may have different weights.)
}
