% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_bugs_regressor_Eti}
\alias{create_bugs_regressor_Eti}
\title{create regressor summation equation used in regression for etiology}
\usage{
create_bugs_regressor_Eti(
  n,
  dm_nm = "dm_Eti",
  b_nm = "betaEti",
  ind_nm = "j",
  sub_ind_nm = "k"
)
}
\arguments{
\item{n}{the length of coefficients}

\item{dm_nm}{name of design matrix; default \code{"dm_Eti"}}

\item{b_nm}{name of the coefficients; default \code{"betaEti"}}

\item{ind_nm}{name of the coefficient iterator; default \code{"j"}}

\item{sub_ind_nm}{name of the subject iterator; default \code{"k"}}
}
\value{
a character string with linear product form
}
\description{
\code{create_bugs_regressor_Eti} creates linear product of coefficients
and a row of design matrix used in regression
}
