% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{create_bugs_regressor_FPR}
\alias{create_bugs_regressor_FPR}
\title{create regressor summation equation used in regression for FPR}
\usage{
create_bugs_regressor_FPR(n, dm_nm = "dm_FPR", b_nm = "b", ind_nm = "j",
  sub_ind_nm = "k")
}
\arguments{
\item{n}{the length of coefficients}

\item{dm_nm}{name of design matrix; default \code{"dm_FPR"}}

\item{b_nm}{name of the coefficients; defaul \code{"b"}}

\item{ind_nm}{name of the coefficient iterator; default \code{"j"}}

\item{sub_ind_nm}{name of the subject iterator; default \code{"k"}}
}
\value{
a character string with linear product form
}
\description{
\code{create_bugs_regressor_FPR} creates linear product of coefficients
and a row of design matrix used in regression
}

