% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{dm_Rdate_Eti}
\alias{dm_Rdate_Eti}
\title{Make etiology design matrix for dates with R format.}
\usage{
dm_Rdate_Eti(Rdate, Y, num_knots_Eti, basis_Eti = "ncs")
}
\arguments{
\item{Rdate}{a vector of dates of R format}

\item{Y}{binary case/control status; 1 for case; 0 for controls}

\item{num_knots_Eti}{number of knots for etiology regression}

\item{basis_Eti}{the type of basis functions to use for etiology regression. It can be "ncs" (natural
cubic splines) or "tprs" (thin-plate regression splines). Default is "ncs".}
}
\value{
Design matrix for Eti regression:
\itemize{\
\item \code{Z_Eti} transformed design matrix for etiology regression
}
}
\description{
\code{dm_Rdate_Eti} creates desigm matrices for etiology regressions.
}

