% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{dm_Rdate_FPR}
\alias{dm_Rdate_FPR}
\title{Make FPR design matrix for dates with R format.}
\usage{
dm_Rdate_FPR(Rdate, Y, num_knots_FPR)
}
\arguments{
\item{Rdate}{a vector of dates of R format}

\item{Y}{binary case/control status; 1 for case; 0 for controls}

\item{num_knots_FPR}{number of knots for FPR regression}
}
\value{
Design matrix for FPR regression:
\itemize{\
\item \code{Z_FPR_ctrl} transformed design matrix for FPR regression for controls
\item \code{Z_FPR_case} transformed design matrix for borrowing FPR
regression from controls to cases. It is obtained using control-standardation,
and square-root matrix of control's $Omega=(abs(outer(knots,knots,"-")))^3$.
}
}
\description{
\code{dm_Rdate_FPR} creates desigm matrices for false positive rate regressions.
}

