% Generated by roxygen2 (4.0.2): do not edit by hand
\name{eda}
\alias{eda}
\title{Exploratory Data Analysis (EDA)}
\usage{
eda(Mobs, Y, X, eda_options, pathogen_display, pathogen_BrS)
}
\arguments{
\item{Mobs}{A list of measurements. The elements of the list should
include \code{MBS}, \code{MSS}, and \code{MGS}.
\itemize{
\item \code{MBS} is a data frame of bronze-standard (BrS) measurements.
Rows are subjects, columns are pathogens targeted in the BrS measurements
(e.g. nasalpharyngeal PCR). These mesaurements have imperfect
sensitivity/specificty.
\item \code{MSS} is a data frame of silver-standard (SS) measurements. Rows are subjects,
  columns are pathogens targeted in the SS measurements (e.g. blood culture).
  These measurements have perfect specificity but imperfect sensitivity.
\item \code{MGS} is a data frame of gold-standard (GS) measurements. Rows are subject,
columns are pathogens targeted in the GS measurements. These measurements
have both perfect sensitivity and perfect specificity.
}}

\item{Y}{A vector of binary variables specifying the disease status (1 for case;
0 for control).}

\item{X}{A design matrix. For false positive rate (FPR) and etiology regressions.}

\item{eda_options}{A list of options for EDA.}

\item{pathogen_display}{A list of pathogen names whose order is to be displayed
for visualizing results.}

\item{pathogen_BrS}{A list of pathogen names which have BrS measurements.}
}
\value{
Tables and figures
}
\description{
Outputs tables and figures to visualize data. Current functionality includes:
\itemize{
\item 1.total no. of positives in cases and controls
\item 2.bubble plot of pairwise associations BrS measurements, separately for controls
  and cases.
}
}

