% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/extract_data_raw.R
\name{extract_data_raw}
\alias{extract_data_raw}
\title{Convert raw data sheet to analyzable format}
\usage{
extract_data_raw(Pathogen, Specimen, Test, X, Xval, MeasDir,
  extra_covariates = NULL, silent = TRUE, individual = NULL)
}
\arguments{
\item{Pathogen}{The vector of pathogen names. It will be reordered according
according to pathogen_type (alphabetical order, "B"->"F"->"V")}

\item{Specimen}{The vector of specimen names}

\item{Test}{The vector of test names}

\item{X}{The vector of covariate names to separately extract data.
For example, in PERCH data cleaning, \code{X = c("newSITE","CASECONT")}}

\item{Xval}{The list of covariate values to stratify data.
Each element corresponds to that in \code{X}. For example, in PERCH
data cleaning, \code{Xval = list("02GAM","1")}}

\item{MeasDir}{The directory to the data set (.csv)}

\item{extra_covariates}{The vector of covariate name for regression purposes.
The default is NULL, which means no such covariate is necessary.}

\item{silent}{Default is \code{TRUE}: the function will not print anything on
the screen; otherwise specify as \code{FALSE}.}

\item{individual}{index for an individual to print his/her measurements}
}
\value{
A list of data. Each element is either a measurement matrix, or
a vector of covariate values
}
\description{
Convert raw data sheet to analyzable format
}

