% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot-etiology.R
\name{get_cause_by_num}
\alias{get_cause_by_num}
\title{get the names of causes by the number of pathogens associated with that cause.
For now it is specifically for PERCH.}
\usage{
get_cause_by_num(cause_list, num)
}
\arguments{
\item{cause_list}{see \code{model_options}}

\item{num}{the number of pathogens for a cause}
}
\value{
a subvector of cause_list
}
\description{
get the names of causes by the number of pathogens associated with that cause.
For now it is specifically for PERCH.
}
\examples{
x <- c("A","B","C","A+B","C+D","AA+BD","A+B+C","HELLO+BYE")
num <- 2
get_cause_by_num(x,num)
}

