% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-etiology-side-by-side.R
\name{get_cause_by_taxo_group}
\alias{get_cause_by_taxo_group}
\title{pick categories from a vector of causes}
\usage{
get_cause_by_taxo_group(cause_list, taxo_group, lookup_table)
}
\arguments{
\item{cause_list}{see \code{model_options}}

\item{taxo_group}{a character string, can be "virus", "bacterium", "combo".}

\item{lookup_table}{a data frame with two columns ("Category" and "Cause").}
}
\value{
a vector of indices indicating if each element belongs to 
\code{taxo_group}.
}
\description{
pick categories from a vector of causes
}

