% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_coverage}
\alias{get_coverage}
\title{Obtain coverage status from a result folder}
\usage{
get_coverage(DIR_NPLCM, truth)
}
\arguments{
\item{DIR_NPLCM}{Path to where Bayesian results are stored}

\item{truth}{True etiologic fraction vector (must sum to 1)  used to generate data.}
}
\value{
A logic vector of length as \code{truth}. 1 for covered; 0 for not.
}
\description{
Obtain coverage status from a result folder
}
