% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot-BrS-panel.R
\name{get_fitted_mean_no_nested}
\alias{get_fitted_mean_no_nested}
\title{get model fitted mean for conditional independence model}
\usage{
get_fitted_mean_no_nested(slice, res_nplcm, model_options, data_nplcm,
  clean_options)
}
\arguments{
\item{slice}{the slice of BrS data that are modeled}

\item{res_nplcm}{matrix of MCMC samples}

\item{model_options}{see \code{\link{nplcm}}}

\item{data_nplcm}{see \code{\link{nplcm}}}

\item{clean_options}{see \code{\link{clean_perch_data}}}
}
\value{
a list with model fitted means
}
\description{
get model fitted mean for conditional independence model
}

