% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nplcm-read-folder.R
\name{get_individual_data}
\alias{get_individual_data}
\title{get individual data}
\usage{
get_individual_data(i, data_nplcm)
}
\arguments{
\item{i}{index of individual as appeared in \code{data_nplcm}}

\item{data_nplcm}{the data for nplcm; see \code{\link{nplcm}}}
}
\value{
a list of the same structure as \code{data_nplcm}; just with one row of values
}
\description{
get individual data
}

