% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/nplcm-read-folder.R
\name{get_individual_prediction}
\alias{get_individual_prediction}
\title{get individual prediction (Bayesian posterior)}
\usage{
get_individual_prediction(read_res)
}
\arguments{
\item{read_res}{a list read from a folder that stores fitted model results;
Commonly, it is the returned value from function \code{\link{nplcm_read_folder}}}
}
\value{
a matrix of individual predictions; rows for cases, columns for causes
specified in \code{model_options$likelihood$cause_list}; See \code{\link{nplcm}}
}
\description{
get individual prediction (Bayesian posterior)
}
\examples{
\dontrun{
get_individual_prediction(nplcm_read_folder("C://2015_09_02_02GAM"))
}
}

