% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_latent_seq}
\alias{get_latent_seq}
\title{get index of latent status}
\usage{
get_latent_seq(cause_list, ord, select_latent = NULL, exact = TRUE)
}
\arguments{
\item{cause_list}{see mode_options in \link{nplcm}}

\item{ord}{order of cause_list according to posterior mean}

\item{select_latent}{Defaut is NULL}

\item{exact}{Default is TRUE}
}
\value{
a vector of indices
}
\description{
get index of latent status
}
