% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot-BrS-panel.R
\name{get_marginal_rates_nested}
\alias{get_marginal_rates_nested}
\title{get marginal TPR and FPR for nested model}
\usage{
get_marginal_rates_nested(slice, res_nplcm, model_options, data_nplcm)
}
\arguments{
\item{slice}{the slice of BrS data that are modeled}

\item{res_nplcm}{matrix of MCMC samples}

\item{model_options}{see \code{\link{nplcm}}}

\item{data_nplcm}{see \code{\link{nplcm}}}
}
\value{
a matrix of no. of rows equal to retained MCMC samples, no. of columns
equal to the no. of measurement dimensions within a slice.
}
\description{
get marginal TPR and FPR for nested model
}

