% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/get_mse.R
\name{get_mse}
\alias{get_mse}
\title{Obtain MSE, bias-squared and variance of posterior from nplcm fitted folders.}
\usage{
get_mse(DIR_NPLCM, truth = NULL, check.truth = FALSE)
}
\arguments{
\item{DIR_NPLCM}{File path where Bayesian results are stored}

\item{truth}{Default is \code{NULL}. True etiologic fraction vector (must sum to 1) used to generate data}

\item{check.truth}{Default is \code{FALSE}. If \code{truth} is a non-null vector,
it asks the function to check the user specified truth against the one used
in the actual simulation.}
}
\value{
a vector of (MSE, bias-squared, variance,truth)
}
\description{
The result is equivalent to Euclidean-type calculation after the
compositional vector (e.g., etiologic fraction) is centered-logratio tranformed.
For simulation only.
}

