% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot-panels.R
\name{get_plot_num}
\alias{get_plot_num}
\title{get the plotting positions (numeric) for the fitted means; 3 positions for each cell}
\usage{
get_plot_num(e, height)
}
\arguments{
\item{e}{Integer index from 1 to length(cause_list)}

\item{height}{the total number of causes}
}
\value{
a triple with numerical plotting positions
}
\description{
get the plotting positions (numeric) for the fitted means; 3 positions for each cell
}

