% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_postsd}
\alias{get_postsd}
\title{Obtain posterior standard deviation from a result folder}
\usage{
get_postsd(DIR_NPLCM)
}
\arguments{
\item{DIR_NPLCM}{Path to where Bayesian results are stored}
}
\value{
a vector of positive numbers
}
\description{
Obtain posterior standard deviation from a result folder
}
