% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{getPckg}
\alias{getPckg}
\title{Get package from CRAN website}
\usage{
getPckg(pckg)
}
\arguments{
\item{pckg}{package name}
}
\value{
None
}
\description{
Get package from CRAN website
}

