% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{H}
\alias{H}
\title{Shannon entropy for multivariate discrete data}
\usage{
H(px)
}
\arguments{
\item{px}{a vector of positive numbers sum to 1}
}
\value{
a non-negative number
}
\description{
Shannon entropy for multivariate discrete data
}

