% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{has_non_basis}
\alias{has_non_basis}
\title{test if a formula has other terms not created by s_date_Eti or s_date_FPR}
\usage{
has_non_basis(form)
}
\arguments{
\item{form}{a formula}
}
\description{
test if a formula has other terms not created by s_date_Eti or s_date_FPR
}
\examples{
form1 <- as.formula(~ -1+s_date_FPR(DATE,Y,basis = "ps",10) + as.factor(SITE))
form2 <- as.formula(~ -1+s_date_FPR(DATE,Y,basis = "ps",10))
form3 <- as.formula(~ s_date_FPR(DATE,Y,basis = "ps",10))

has_non_basis(form1)
has_non_basis(form2)
has_non_basis(form3)
}
