% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{I2symb}
\alias{I2symb}
\title{Convert 0/1 coding to pathogen/combinations}
\usage{
I2symb(binary_code, pathogen_list)
}
\arguments{
\item{binary_code}{Binary indictors for pathogens}

\item{pathogen_list}{The complete list of pathogen names}
}
\value{
The name of pathogen or pathogen combination indicated by "code"
}
\description{
Reverse to \code{\link{symb2I}}
}
\examples{
I2symb("001",c("A","B","C"))
I2symb("000",c("A","B","C"))
}

