% Generated by roxygen2 (4.0.2): do not edit by hand
\name{Imat2cat}
\alias{Imat2cat}
\title{Convert a matrix of binary indicators to categorial variables}
\usage{
Imat2cat(binary_mat, allowed_list, pathogen_list)
}
\arguments{
\item{binary_mat}{The matrix of binary indicators. Rows for subjects, columns for pathogens in the \code{"pathogen.list"}}

\item{allowed_list}{The list of allowed single pathogen or pathogen combination}

\item{pathogen_list}{The complete list of pathogen names}
}
\value{
A vector of categorical variables. Its length equals the length of \code{"allowed.list"}
}
\description{
Convert a matrix of binary indicators to categorial variables
}
\examples{
Imat2cat(rbind(diag(3),c(1,1,0),c(0,0,0)),c("A","B","C","A+B","NoA"),c("A","B","C"))
}

