% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/insert-bugfile-chunk.R
\name{insert_bugfile_chunk_noreg_etiology}
\alias{insert_bugfile_chunk_noreg_etiology}
\title{insert etiology code chunks into .bug file}
\usage{
insert_bugfile_chunk_noreg_etiology()
}
\value{
a long character string to be inserted into target .bug model file
}
\description{
insert etiology code chunks into .bug file
}

