% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/insert-bugfile-chunk.R
\name{insert_bugfile_chunk_noreg_meas}
\alias{insert_bugfile_chunk_noreg_meas}
\title{insert bronze-standard measurements code chunks into .bug file}
\usage{
insert_bugfile_chunk_noreg_meas(Mobs, cause_list, meas = "BrS")
}
\arguments{
\item{Mobs}{measurement data in the form of \code{data_nplcm}}

\item{cause_list}{a list of latent status (crucial for building templates)}

\item{meas}{"BrS", or "SS"}
}
\value{
a long character string to be inserted into target .bug model file
}
\description{
insert bronze-standard measurements code chunks into .bug file
}

