% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert-bugfile-chunk.R
\name{insert_bugfile_chunk_reg_etiology}
\alias{insert_bugfile_chunk_reg_etiology}
\title{insert etiology regression for latent status code chunk into .bug file}
\usage{
insert_bugfile_chunk_reg_etiology(Eti_formula, ppd = NULL)
}
\arguments{
\item{Eti_formula}{Etiology regression formula; Check \code{model_options$likelihood$Eti_formula}.}

\item{ppd}{Default is NULL; set to TRUE for posterior predictive checking}
}
\value{
a long character string to be inserted into .bug model file 
as distribution specification for latent status
}
\description{
insert etiology regression for latent status code chunk into .bug file
}

