% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert-bugfile-chunk.R
\name{insert_bugfile_chunk_reg_nonest_meas}
\alias{insert_bugfile_chunk_reg_nonest_meas}
\title{Insert measurement likelihood (with regression) code chunks into .bug model file}
\usage{
insert_bugfile_chunk_reg_nonest_meas(Mobs, prior, cause_list, FPR_formula,
  use_measurements = "BrS", ppd = NULL, use_jags = FALSE)
}
\arguments{
\item{Mobs}{Measurement data in the form of \code{data_nplcm}}

\item{prior}{Prior specification from \code{model_options}}

\item{cause_list}{A list of latent status names (crucial for building templates; 
see \code{\link{make_template}})}

\item{FPR_formula}{A list of FPR regression formula; check \code{model_options$likelihood$FPR_formula}}

\item{use_measurements}{"BrS", or "SS"}

\item{ppd}{Default is NULL; set to TRUE for posterior predictive checking}

\item{use_jags}{Default is FALSE; set to TRUE if want to use JAGS for model fitting.}
}
\value{
A long character string to be inserted into .bug model file as measurement
likelihood
}
\description{
Insert measurement likelihood (with regression) code chunks into .bug model file
}
\seealso{
It is used in \link{write_model_Reg_NoNest} for constructing a .bug file along with
specification of latent status regression (\link{insert_bugfile_chunk_reg_etiology})
}

