% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{is.error}
\alias{is.error}
\title{Test for 'try-error' class}
\usage{
is.error(x)
}
\arguments{
\item{x}{An object to be test if it is "try-error"}
}
\value{
Logical. \code{TRUE} for "try-error"; \code{FALSE} otherwise
}
\description{
Test for 'try-error' class
}
\references{
\url{http://adv-r.had.co.nz/Exceptions-Debugging.html}
}

