% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_intercept_only}
\alias{is_intercept_only}
\title{outputs logical values for a formula; to identify intercept-only formula.}
\usage{
is_intercept_only(form)
}
\arguments{
\item{form}{Regression formula}
}
\value{
\code{TRUE} for intercept-only; \code{FALSE} otherwise
}
\description{
outputs logical values for a formula; to identify intercept-only formula.
}

