% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nplcm-read-folder.R
\name{is_jags_folder}
\alias{is_jags_folder}
\title{See if a result folder is obtained by JAGS}
\usage{
is_jags_folder(DIR_NPLCM)
}
\arguments{
\item{DIR_NPLCM}{directory to the folder with results. 
"mcmc_options.txt" must be in the folder.}
}
\value{
TRUE for from JAGS; FALSE otherwise.
}
\description{
See if a result folder is obtained by JAGS
}
