% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot-panels.R
\name{is_length_all_one}
\alias{is_length_all_one}
\title{check if a list has elements all of length one}
\usage{
is_length_all_one(x)
}
\arguments{
\item{x}{a list}
}
\value{
TRUE or FALSE
}
\description{
check if a list has elements all of length one
}

