% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{load_or_install}
\alias{load_or_install}
\title{Load or install a package}
\usage{
load_or_install(package_names, repos = "http://lib.stat.cmu.edu/R/CRAN")
}
\arguments{
\item{package_names}{A vector of package names}

\item{repos}{URL for downloading the packages. Default is
\url{http://lib.stat.cmu.edu/R/CRAN}}
}
\value{
No message if it successfully loads the specified packages; Error
 if such a package does not exist.
}
\description{
\code{load_or_install} checks if a package is installed,
loads it if it is, and installs it if not.
}
\references{
Credit:
 \url{http://www.vikparuchuri.com/blog/loading-andor-installing-packages/}
}
