% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/logORmat.R
\name{logORmat}
\alias{logORmat}
\title{Visualize pairwise log odds ratios (LOR) for BrS data}
\usage{
logORmat(MBS.case, MBS.ctrl, pathogen_display, pathogen_BrS,
  logOR_rounding = 2)
}
\arguments{
\item{MBS.case}{Case BrS measurements.}

\item{MBS.ctrl}{Control BrS measurements.}

\item{pathogen_display}{The pathogen vector in desired order for display.
It can be of larger length than that of \code{pathogen_BrS}.}

\item{pathogen_BrS}{The vector of pathogen names corresponding to each
row/column in the plot. All pathogens have BrS measures.}

\item{logOR_rounding}{Rounding number of the log odds ratio. Default is 2.}
}
\value{
Figure of LOR matrix and relavent s.e. and significance information.
}
\description{
The matrix of values is with controls on upper right and cases on lower
left. Log odds ratio (LOR) is at the top of the cell.  Below it, its standard
error is in smaller type, using the same color as the LOR.  Then the
estimate is divided by its standard error. If it is less than 1 in
absolute value, we put a plus (red) or minus (blue) when
the Z-stat is between 1-2 in absolute value and put the actual
value when the Z is greater than 2.
}

