% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{logsumexp}
\alias{logsumexp}
\title{log sum exp trick}
\usage{
logsumexp(x)
}
\arguments{
\item{x}{a vector of numbers}
}
\value{
a numeric value
}
\description{
log sum exp trick
}
\examples{

logsumexp(c(-20,-30))

}
