% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{lookup_quality}
\alias{lookup_quality}
\title{get position to store in data_nplcm$Mobs:}
\usage{
lookup_quality(quality_nm)
}
\arguments{
\item{quality_nm}{names of quality: can be "BrS", "SS" or "GS"}
}
\value{
position of the quality name: "BrS"-1; "SS"-2; "GS"-3.
}
\description{
get position to store in data_nplcm$Mobs:
}
\details{
also works for a vector
}
\examples{
\dontrun{
lookup_quality("BrS")
lookup_quality("HH")
}
}

