% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot-etiology.R
\name{lookup_side_by_side}
\alias{lookup_side_by_side}
\title{a function to match each model's result to the desired order:}
\usage{
lookup_side_by_side(disp_order, union_nm, nm_list)
}
\arguments{
\item{disp_order}{the desired order}

\item{union_nm}{the union of all cause names}

\item{nm_list}{the list of names for each model (i.e. folder)}
}
\value{
a list of two elements:
\itemize{
\item \code{model_res} a list of length same as \code{DIR_list} in
\code{\link{plot_etiology_side_by_side}}. Each element is the order
for that particular name in the list; the length is the same as union_nm.
\item \code{union_res} the names of the causes after union and re-ordering
}
}
\description{
a function to match each model's result to the desired order:
}
\examples{
disp_order <- c("B","E","D","C","F","A")
  union_nm   <- c("A","B","C","D","E")
  nm_list <-list(c("C","E"),
                 c("D","B"),
                 c("C","A","E"))
  lookup_side_by_side(disp_order,union_nm,nm_list)
}

