% Generated by roxygen2 (4.0.2): do not edit by hand
\name{make_foldername}
\alias{make_foldername}
\title{Create new folder name}
\usage{
make_foldername(parent_path, parameter_names, parameter_vals)
}
\arguments{
\item{parent_path}{The parent directory where to put the new folder}

\item{parameter_names}{The parameters that distinguish this folder's scenario}

\item{parameter_vals}{The actual parameter values}
}
\value{
A string for folder name
}
\description{
Create new folder name
}

