% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{make_numbered_list}
\alias{make_numbered_list}
\title{Make a list with numbered names}
\usage{
make_numbered_list(...)
}
\arguments{
\item{...}{any R object}
}
\value{
a list with names numbered
}
\description{
To collect multiple measurements within the same category, e.g., bronze-standard.
}

