% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{make_template}
\alias{make_template}
\title{make template for fitting nplcm model}
\usage{
make_template(cause_list, pathogen_BrS, pathogen_BrS_combined = NULL,
  combined_causes2meas = NULL)
}
\arguments{
\item{pathogen_BrS}{a vector of strings of all possible pathogen measurements. If
there are multiple BrS measurements, this vector should be the one with the finest
labeling of pathogens}

\item{pathogen_BrS_combined}{Default is \code{NULL}. It is a vector of strings of extra measurements. It maybe
consisted of some exact names from \code{pathogen_BrS} or it has a new name that combine
some names in \code{pathogen_BrS}. See example.}

\item{combined_causes2meas}{Default is \code{NULL}. It is a list with two elements. The first
is the indices of measurments to combine in the finer vector \code{pathogen_BrS}; the second element
is the position in \code{pathogen_BrS_combined} that represents the combined measurement names. It currently
assumes that the two elements have the same first index. See example.}

\item{cause_lista}{a vector of strings representing latent status}
}
\value{
a matrix with rows being latent status, columns being measured biomarkers (pathogens)
}
\description{
template is a function of a cause list and measurement list
}
\examples{
cause_list <- c("A","B1","B2","C","D","E")
pathogen_BrS_1 <- c("A","B1","B2","C","D","E")
pathogen_BrS_2 <- c("A","B","C","D","E")
combined_causes2meas <- list(comb_cause = c(2,3),comb_meas = 2)
make_template(cause_list,pathogen_BrS_1)
make_template(cause_list,pathogen_BrS_1, pathogen_BrS_2,combined_causes2meas)
}

