% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{marg_H}
\alias{marg_H}
\title{Shannon entropy for binary data}
\usage{
marg_H(m_px)
}
\arguments{
\item{m_px}{a number between 0 and 1}
}
\value{
a non-negative number
}
\description{
Shannon entropy for binary data
}
\examples{

marg_H(0.1)

}
