% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{merge_lists}
\alias{merge_lists}
\title{For a list of many sublists each of which has matrices as its member, 
we combine across the many sublists to produce a final list}
\usage{
merge_lists(list_of_lists)
}
\arguments{
\item{list_of_lists}{a list of sublists}
}
\value{
a list after merge
}
\description{
For a list of many sublists each of which has matrices as its member, 
we combine across the many sublists to produce a final list
}
\examples{
DT1 = list(A=1:3,B=letters[1:3])
DT2 = list(A=4:5,B=letters[4:5])
DT3 = list(A=1:4,B=letters[1:4])
DT4 = list(A=4:7,B=letters[4:7])
l = list(DT1,DT2);names(l) <- c("haha","hihi")
l2 = list(DT3,DT4);names(l2) <- c("haha","hihi")
listoflists <- list(l,l2);names(listoflists) <- c("dude1","dude2")
listoflists
merge_lists(listoflists)
}
\seealso{
Other data operation functions: 
\code{\link{combine_data_nplcm}()},
\code{\link{subset_data_nplcm_by_index}()}
}
\concept{data operation functions}
