% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/nplcm_checking_common_pattern_two_folders.R
\name{nplcm_checking_common_pattern_two_folders}
\alias{nplcm_checking_common_pattern_two_folders}
\title{Posterior predictive checking for the nested partially class models -
frequent patterns in the BrS data.}
\usage{
nplcm_checking_common_pattern_two_folders(DIR_NPLCM1, DIR_NPLCM2,
  npat.case = 10, npat.ctrl = 10, cex.pattern = 1)
}
\arguments{
\item{DIR_NPLCM1}{File path 1 to the folder that stores results from npLCM fit.}

\item{DIR_NPLCM2}{File path 2 to the folder that stores results from npLCM fit.}

\item{npat.case}{Number of the most common BrS measurement pattern among cases.
Default is 10.}

\item{npat.ctrl}{Number of the most common BrS measurement pattern among controls.
Default is 10.}

\item{cex.pattern}{Size of patterns; Default is 1.}
}
\value{
A figure of posterior predicted frequencies compared with the observed
frequencies of the most common patterns for the BrS data. The function generates
this figure in your working directory automatically.
}
\description{
At each MCMC iteration, we generate a new data set based on the model and
parameter values at that iteration. The sample size of the new data set equals
that of the actual data set, i.e. the same number of cases and controls.
}

