% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/nplcm_fit_NoReg_BrS_NoNest.R
\name{nplcm_fit_NoReg_BrS_NoNest}
\alias{nplcm_fit_NoReg_BrS_NoNest}
\title{Fit nested partially-latent class model (low-level)}
\usage{
nplcm_fit_NoReg_BrS_NoNest(Mobs, Y, X, model_options, mcmc_options)
}
\arguments{
\item{Mobs}{A list of measurements. The elements of the list
should include \code{MBS}, \code{MSS}, and \code{MGS}.
\itemize{
\item \code{MBS} a data frame of bronze-standard (BrS) measurements.
Rows are subjects, columns are pathogens.
They have imperfect sensitivity/specificity.
(e.g. nasalpharyngeal PCR). These mesaurements have imperfect sensitivity/specificty.
\item \code{MSS} a data frame of silver-standard (SS) measurements.
Rows are subjects, columns are pathogens (e.g. blood culture).
These measurements have perfect specificity but imperfect sensitivity.
\item \code{MGS} a data frame of gold-standard (GS) measurements.
Rows are subject, columns are pathogens.
These measurements have perfect sensitivity and specificity.
}}

\item{Y}{Vector of disease status: 1 for case, 0 for control.}

\item{X}{Design matrix. For regression modeling.}

\item{model_options}{A list of model options.

\itemize{
\item \code{M_use} List of measurements to be used in the model;
\item \code{k_subclass}The number of nested subclasses. 1 for conditional independence,
>1 for conditional dependence;
\item \code{TPR_prior} Description of priors for the measurements
(e.g., informative vs non-informative).
Its length should be the same with \code{M_use};
\item \code{Eti_prior} Description of etiology prior
(e.g., overall uniform - all hyperparameters are 1, v.s. "0_1" - all hyperparameters
are 0.1);
\item \code{pathogen_BrS_list} The vector of pathogen names with BrS measure;
\item \code{cause_list} The vector of causes that are either singleton or
combinations of items in \code{pathogen_BrS_list}; 'NoA' can also be included
at the end, meaning 'None-of-Above';
\item \code{X_reg_FPR} The vector of covariate names that stratify/regress
false positive rates (FPR);
\item \code{X_reg_Eti} The vector of covariate names that stratify/regress
 etiologies;
\item \code{pathogen_cat} The two-column dataframe that has category of pathogens: virus (V), bacteria (B)
and fungi (F);
\item \code{pathogen_SSonly_list} The vector of pathogens with only
SS measure;
\item \code{pathogen_SSonly_cat} The category of pathogens with only SS measure.
}}

\item{mcmc_options}{A list of Markov chain Monte Carlo (MCMC) options.

\itemize{
\item \code{debugstatus} for whether to pause WinBUGS after it finishes
model fitting;
\item \code{n.chains} for the number of MCMC chains;
\item \code{n.burnin} for the number of burn-in samples;
\item \code{n.thin} keep every other n.thin samples after burn-in period;
\item \code{individual.pred} whether to perform individual prediction;
\item \code{ppd} whether to perform posterior predictive (ppd) checking;
\item \code{result.folder} for the path to folder storing the results;
\item \code{bugsmodel.dir} for the directory to WinBUGS model files;
\item \code{winbugs.dir} for the directory where WinBUGS 1.4 is installed.
}}
}
\value{
WinBUGS fit results.
}
\description{
Features:
\itemize{
\item no regression;
\item bronze- (BrS) measurements;
\item conditional independence;
\item all pathogens have BrS measurements.
}
}

