% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/nplcm_plot_BrS_panel.R
\name{nplcm_plot_BrS_panel}
\alias{nplcm_plot_BrS_panel}
\title{Plot bronze-standard (BrS) panel}
\usage{
nplcm_plot_BrS_panel(MBS, model_options, clean_options, res_nplcm, bugs.dat,
  top_BrS = 1.3, prior_shape = "boxplot", cexval = 1, srtval = 0)
}
\arguments{
\item{MBS}{Matrix of bronze-standard measurements. Rows for subjects (cases at
the top, controls at the bottom), columns for pathogen-specimen combination.}

\item{model_options}{See \code{\link{nplcm}}}

\item{clean_options}{See \code{\link{perch_data_clean}}}

\item{res_nplcm}{See \code{\link{nplcm_read_folder}}}

\item{bugs.dat}{Data input for the model fitting.}

\item{top_BrS}{Numerical value to specify the rightmost limit
on the horizontal axis for the BrS panel.}

\item{prior_shape}{\code{interval} or \code{boxplot} - for how to represent
prior/posteriors of the TPR/FPRs of measurements.}

\item{cexval}{Default is 1 - size of text of the BrS percentages.}

\item{srtval}{Default is 0 - the direction of the text for the BrS percentages.}
}
\description{
Now only works for singleton etiologies.
}

