% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/nplcm_plot_etiology_music_sheet.R
\name{nplcm_plot_etiology_music_sheet}
\alias{nplcm_plot_etiology_music_sheet}
\title{compare the posterior distribution of population etiologies side-by-side}
\usage{
nplcm_plot_etiology_music_sheet(DIR_list, DIR_pathogen_displayorder_lookup,
  dir_separator = "[\\\\]")
}
\arguments{
\item{DIR_list}{The list of directory paths, each storing a model output.}

\item{DIR_pathogen_displayorder_lookup}{The directory path to the .csv file
that stores the display order of pathogens in the combined music sheet plot.}

\item{dir_separator}{The separator character for a file path.}
}
\value{
A figure that compares posterior etiology distribution stored in
two or more folders
}
\description{
The so-called "music-sheet plot". It takes two, or three or more output
folders so that the results are combined. There must be a check to make sure
that pathogens from one analysis are subset of another. DN: current implementation
does not check whether the results are from the same site, or strata. This
information can be seen from the folder names, or the figure names within
 the folders.
}

