% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/nplcm_plot_individual_diagnosis.R
\name{nplcm_plot_individual_diagnosis}
\alias{nplcm_plot_individual_diagnosis}
\title{Individual prediction from nested partially-latent class models}
\usage{
nplcm_plot_individual_diagnosis(DIR_NPLCM, npat = 16)
}
\arguments{
\item{DIR_NPLCM}{The file path to the folder that stores results from npLCM fit.}

\item{npat}{The number of most frequent patterns among cases to display.}
}
\value{
A figure with individual diagnosis plots. The bronze-standard (BrS) measurement pattern is listed
at the top of the barplots. The percentage below it is the observed frequency in the actual data.
Another 1/0 pattern is shown in red, and represents the silver-standard(SS) measurements.
The percentage below it means, within cases having the BrS
pattern, the fraction of cases having this SS pattern.
}
\description{
DN:1.currently only works for both BrS and SS data, and SSonly data. Any causes.
}

