% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/nplcm_plot_individual_diagnosis_two_folders.R
\name{nplcm_plot_individual_diagnosis_two_folders}
\alias{nplcm_plot_individual_diagnosis_two_folders}
\title{Individual prediction from nested partially-latent class models (comparing
posterior results stored in two folders; expecting the fits are for the same
data set.)}
\usage{
nplcm_plot_individual_diagnosis_two_folders(DIR_NPLCM1, DIR_NPLCM2, npat = 16)
}
\arguments{
\item{DIR_NPLCM1}{The first file path to the folder that stores results from npLCM fit.}

\item{DIR_NPLCM2}{}

\item{npat}{The number of most frequent patterns among cases to display.}
}
\value{
A figure with individual diagnosis plots. The Bronze-Standard (BrS)
 measurement pattern is listed at the top of the barplots. Below it is the
 percentage of the observed frequency of that pattern among cases.
 A second vector of 1/0 pattern is shown in red. It represents the
Silver-Standard(SS) measurements. The percentage below it is the fraction of
 cases having this SS pattern among cases with the BrS pattern on the top.
 The whole list of causes are shown at the bottom.
}
\description{
DN:1.currently only works for 1) both BrS and SS data, and SS only data; 2)
both BrS and SS data, without SSonly data.
The cause list can be of any causes.
}

