% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/nplcm_plot_skyline_plot.R
\name{nplcm_plot_skyline}
\alias{nplcm_plot_skyline}
\title{Distribution of individual prediction for each pathogen sort by specified order}
\usage{
nplcm_plot_skyline(DIR_NPLCM, DIR_pathogen_displayorder_lookup)
}
\arguments{
\item{DIR_NPLCM}{The file path to the folder that stores results from npLCM fit.}

\item{DIR_pathogen_displayorder_lookup}{The directory path to the .csv file
that stores the display order of pathogens in the combined music sheet plot.}
}
\value{
A list of histograms of individual prediction for each pathogen, overall population
etiolgy are put in the top-middle section with abbreviated pathogen name; etiology
contributed by each quartile are put at the corresponding section with proportion
contributed to overall etiology by each quartile listed below with brackets. Black histograms
are for subset of kids with that particular pathogen detected in NP swab.
}
\description{
The so-called "skyline-plot" are histograms of individual prediction for each pathogen.
}

