% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/nplcm_read_folder.R
\name{nplcm_read_folder}
\alias{nplcm_read_folder}
\title{Read data and other model information from a folder that stores model results.}
\usage{
nplcm_read_folder(DIR_NPLCM)
}
\arguments{
\item{DIR_NPLCM}{File path to the folder containing posterior samples}
}
\value{
None
}
\description{
Read data and other model information from a folder that stores model results.
}

