% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{null_as_zero}
\alias{null_as_zero}
\title{Convert \code{NULL} to zero.}
\usage{
null_as_zero(x)
}
\arguments{
\item{x}{A number (usually a member of a list) that might be \code{NULL}}
}
\value{
A number
}
\description{
\code{null_as_zero} make \code{NULL} to be zero.
}

