% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-panels.R
\name{order_post_eti}
\alias{order_post_eti}
\title{order latent status by posterior mean}
\usage{
order_post_eti(res_nplcm, model_options)
}
\arguments{
\item{res_nplcm}{result from model fits}

\item{model_options}{model specification}
}
\value{
a list with order (\code{ord}) and ordered posterior samples (by column)
}
\description{
order latent status by posterior mean
}

