% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set-prior.R
\name{overall_uniform}
\alias{overall_uniform}
\title{specify overall uniform (symmetric Dirichlet distribution) for etiology prior}
\usage{
overall_uniform(alpha, cause_list)
}
\arguments{
\item{alpha}{any positive number, usually 1.}

\item{cause_list}{a list of latent status}
}
\value{
a vector of length \code{length(cause_list)}
}
\description{
specify overall uniform (symmetric Dirichlet distribution) for etiology prior
}
\seealso{
Other prior specification functions: 
\code{\link{set_prior_tpr_BrS_NoNest}()},
\code{\link{set_prior_tpr_SS}()}
}
\concept{prior specification functions}
