% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gateaux_job_runner.R
\name{parse_joblist}
\alias{parse_joblist}
\title{Call arbitrary gateaux jobs from R}
\usage{
parse_joblist(json_ret, prefix, append = T)
}
\arguments{
\item{json_ret}{return string from gateaux API}

\item{prefix}{for the list, can include path}

\item{append}{append to existing file?}
}
\description{
Call arbitrary gateaux jobs from R
}
\author{
Dragonfly bakery
}
