% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\docType{data}
\name{pathogen_category_perch}
\alias{pathogen_category_perch}
\title{pathogens and their categories in PERCH study (virus or bacteria)}
\format{
A matrix of two columns
\describe{
\item{pathogen}{names of the pathogens}
\item{pathogen_type}{category of the pathogens, \code{B} for
bacterium, \code{V} for virus, \code{F} for fungus, \code{O} for "not categorized"}
}
}
\usage{
data("pathogen_category_perch")
}
\value{
No returned value; just loading data into the working space.
}
\description{
231 rows indicating bacteria, virus, fungi, or other categories.
}
\keyword{datasets}
