% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\docType{data}
\name{pathogen_category_simulation}
\alias{pathogen_category_simulation}
\title{Hypothetical pathogens and their categories (virus or bacteria)}
\format{
A matrix of two columns
\describe{
  \item{pathogen}{names of the hypothetical pathogens, A-Z}
  \item{pathogen_type}{category of the hypothetical pathogens, \code{B} for
  bacterium, \code{V} for virus, which are randomly assigned.}
}
}
\usage{
pathogen_category_simulation
}
\description{
This is used in simulations where the pathogen names are from the alphabet,
and we hope to plot etiologies grouped by virus or bacteria
}
\examples{
data("pathogen_category_simulation")
str(pathogen_category_simulation)
}
\keyword{datasets}
