% Generated by roxygen2 (4.0.2): do not edit by hand
\name{perch_data_clean}
\alias{perch_data_clean}
\title{Clean PERCH data}
\usage{
perch_data_clean(clean_options)
}
\arguments{
\item{clean_options}{The list of options for cleaning PERCH data.
The specific elements are as follows:
\itemize{
\item{\code{case_def}}{: variable name for case definition;}
\item{\code{case_def_val}}{: The value of the case-definition variable;}
\item{\code{ctrl_def}}{: variable name for control definition;}
\item{\code{ctrl_def_val}}{: value of the control-definition variable;}
\item{\code{X_strat}}{: A vector of strings, each defining the variables used to
stratify the data;}
\item{\code{X_strat_val}}{: A list of actual values that X_strat should take to
get stratified data sets;}
\item{\code{pathogen_BrS}}{: The vector of pathogen names (arbitrary order).
It has to be a subset of pathogen category information in the \code{PathCatDir}.}
\item{\code{X_extra}}{: A vector of strings, each being the covariate name that one
wants to have in the data set to be analyzed;}
\item{\code{X_order_obs}}{: A vector of strings, each being the covariate name that one
wants to order the rows of the observations. This helps observations with same
site to be put together, or helps order dates from earlier to later;}
\item{\code{RawMeasDir}}{: The file path to the raw data set;}
\item{\code{write_newSite}}{: Must set to \code{TRUE} if the raw data set is changed;}
\item{\code{newSite_write_Dir}}{: The file path where a new/cleaned data set for actual
 analyses will be written;}
\item{\code{MeasDir}}{: The file path to the cleaned data set;}
\item{\code{PathCatDir}}{: The file path to the pathogen category list (.csv). This list
should be as complete as possible to display all pathogens used in an actual
analysis;}
\item{\code{allow_missing}}{: TRUE for the model to using an observation that has either
BrS missing, or SS missing, or both missing. It has to be set to \code{TRUE}
in order to get all the BCX information available, albeit some cases have missing
NPPCR data;}
}}
}
\value{
A List: \code{list(Mobs,Y,X,JSS,pathogen_MSS_ordered,pathogen_cat)}, or
with additional \code{JSSonly, pathogen_SSonly_cat} if silver-
standard only pathogens are supplied.
\itemize{
\item \code{Mobs} A list of bronze- (\code{MBS}), silver- (\code{MSS}),
and gold-standard (\code{MGS}, if available) measurements. Here if all
pathogens have BrS measures, MSS has the same number of columns as in MBS;
if some pathogens only have SS measures, then MSS will have extra columns
for these measurements;
\item \code{Y} Binary indicator for cases (1) and controls (0);
\item \code{X} Data frame of covariates for cases and controls. The names
are specified in \code{X_extra};
\item \code{JSS} Number of pathogens with both silver- and bronze-standard
data;
\item \code{pathogen_MSS_ordered} Ordered vector of pathogen names. Pathogens
with both SS and BrS pathogens are ordered first and then those with only BrS
measurements. Pathogens with only silver-standard measures are not included.
\item \code{pathogen_cat} Pathogen categories ordered according to
\code{pathogen_MSS_ordered}.
\item \code{JSSonly} Number of pathogens with only silver-standard data;
\item \code{pathogen_SSonly_cat} Category of pathogens with only silver-standard
data.
}
The function does not order silver-standard data only pathogens.
}
\description{
It deletes cases with two positives for BCX measures. To order cases
and controls separately according to site and enrollment date, we suggest put
\code{c("newSITE","ENRLDATE")} in \code{X_extra} and \code{X_order_obs} as listed
below.
}

