% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-check-common-pattern.R
\name{plot_check_common_pattern}
\alias{plot_check_common_pattern}
\title{Posterior predictive checking for the nested partially class models - 
frequent patterns in the BrS data. (for multiple folders)}
\usage{
plot_check_common_pattern(DIR_list, slice_vec = rep(1, length(DIR_list)),
  n_pat = 10, dodge_val = 0.8)
}
\arguments{
\item{DIR_list}{The list of directory paths, each storing a model output.}

\item{slice_vec}{Default are 1s, for the first slice of BrS data.}

\item{n_pat}{Number of the most common BrS measurement pattern among cases and controls.
Default is 10.}

\item{dodge_val}{Default is 0.8; For width of boxplots.}
}
\value{
A figure of posterior predicted frequencies compared with the observed 
frequencies of the most common patterns for the BrS data.
}
\description{
At each MCMC iteration, we generate a new data set based on the model and 
parameter values at that iteration. The sample size of the new data set equals
that of the actual data set, i.e. the same number of cases and controls.
}
\examples{
\dontrun{
DIR_list <- list("C:\\\\2015_09_04_05SAF_k=1","C:\\\\2015_09_04_05SAF_k=2")
plot_check_common_pattern(DIR_list)
plot_check_common_pattern(DIR_list[[1]])
}

}
\seealso{
Other visualization functions: \code{\link{plot_BrS_panel}},
  \code{\link{plot_SS_panel}},
  \code{\link{plot_check_pairwise_SLORD}},
  \code{\link{plot_etiology_regression}},
  \code{\link{plot_etiology_side_by_side}},
  \code{\link{plot_etiology_strat}},
  \code{\link{plot_group_etiology}},
  \code{\link{plot_panels}}, \code{\link{plot_pie_panel}},
  \code{\link{plot_selected_etiology}}
}
\concept{model generating functions}
\concept{visualization functions}
