% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-etiology-regression.R
\name{plot_etiology_regression}
\alias{plot_etiology_regression}
\title{visualize the etiology regression with a continuous covariate}
\usage{
plot_etiology_regression(DIR_NPLCM, stratum_bool, plot_basis = FALSE,
  truth = NULL)
}
\arguments{
\item{DIR_NPLCM}{File path to the folder containing posterior samples}

\item{stratum_bool}{a vector of TRUE/FALSE with TRUE indicating the rows of subjects to include}

\item{plot_basis}{TRUE for plotting basis functions; Default to FALSE}

\item{truth}{a list of truths computed from true parameters in simulations; elements: Eti, FPR, PR
which are matrices of # of rows = # of subjects, # columns: \code{length(cause_list)} for Eti
and \code{ncol(data_nplcm$Mobs$MBS$MBS1)}; Default to \code{NULL} for real data analyses.}
}
\value{
A figure of etiology regression curves and some marginal positive rate assessment of
model fit; See example for the legends.
}
\description{
This function visualizes the etiology regression against one continuous covariates, e.g., 
enrollment date. (NB: dealing with NoA, multiple-pathogen causes, other continuous covariates?)
}
\examples{
\dontrun{
# legend.text = c("[UPPER FIGURES]",
                "observed prevalence: cases",
                "observed prevalence: controls",
                "fitted prevalence: cases",
                "fitted prevalence: controls",
                "true positive rate: mean",
                "true positive rate: 95\%CI",
                "[BOTTOM FIGURES]",
                "etiology curve: mean",
                "overall etiology: mean",
                "overall etiology: 95\%CI","","","")
legend.col=c("white","black","dodgerblue2","black","dodgerblue2","red","red",
             "white","springgreen4","orange","orange","white","white","white")
legend.lty=c(1,2,2,1,1,1,2,1,1,1,2)
legend.lwd=c(2,2,2,2,2,2,2,2,2,2,2,2,2,2)
legend("topleft",legend=legend.text,
       lty=legend.lty,lwd=legend.lwd,
       col=legend.col,ncol=2,
       y.intersp=1.5,cex=1.6,box.col=NA)
       }
}
\references{
See example figures:
\url{https://github.com/zhenkewu/baker/blob/master/inst/figs/visualize_etiology_regression_SITE=1.pdf}
and its legends: 
\url{https://github.com/zhenkewu/baker/blob/master/inst/figs/legends_visualize_etiology_regression.png}
}
\seealso{
Other visualization functions: \code{\link{plot_BrS_panel}},
  \code{\link{plot_SS_panel}},
  \code{\link{plot_check_common_pattern}},
  \code{\link{plot_check_pairwise_SLORD}},
  \code{\link{plot_etiology_side_by_side}},
  \code{\link{plot_etiology_strat}},
  \code{\link{plot_group_etiology}},
  \code{\link{plot_panels}}, \code{\link{plot_pie_panel}},
  \code{\link{plot_selected_etiology}}
}
\concept{visualization functions}
