% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-etiology-regression.R
\name{plot_etiology_strat}
\alias{plot_etiology_strat}
\title{visualize the etiology estimates for each discrete levels}
\usage{
plot_etiology_strat(DIR_NPLCM, strata_weights = NULL, truth = NULL)
}
\arguments{
\item{DIR_NPLCM}{File path to the folder containing posterior samples}

\item{strata_weights}{a vector of weights that sum to one; for each pathogen
the weights specify how the j-th etiology fraction should be combined across all
levels of the discrete predictors in the data}

\item{truth}{a list of true values, e.g., 
\code{truth=list(allEti = a list of etiology fractions)}}
}
\description{
This function visualizes the etiology estimates against one discrete covariate, e.g., 
age groups.
}
\seealso{
Other visualization functions: \code{\link{plot_BrS_panel}},
  \code{\link{plot_SS_panel}},
  \code{\link{plot_check_common_pattern}},
  \code{\link{plot_check_pairwise_SLORD}},
  \code{\link{plot_etiology_regression}},
  \code{\link{plot_etiology_side_by_side}},
  \code{\link{plot_group_etiology}},
  \code{\link{plot_panels}}, \code{\link{plot_pie_panel}},
  \code{\link{plot_selected_etiology}},
  \code{\link{plot_subwt_regression}}
}
\concept{visualization functions}
