% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-etiology-regression.R
\name{plot_etiology_strat_nested}
\alias{plot_etiology_strat_nested}
\title{visualize the etiology estimates for each discrete levels}
\usage{
plot_etiology_strat_nested(
  DIR_NPLCM,
  strata_weights = NULL,
  truth = NULL,
  RES_NPLCM = NULL
)
}
\arguments{
\item{DIR_NPLCM}{File path to the folder containing posterior samples}

\item{strata_weights}{a vector of weights that sum to one; for each pathogen
the weights specify how the j-th etiology fraction should be combined across all
levels of the discrete predictors in the data; can also specify as \code{"empirical"}
to use empirical weights (fractions of subjects in each stratum)}

\item{truth}{a list of true values, e.g.,
\verb{truth=list(allEti = a list of etiology fractions)}}

\item{RES_NPLCM}{pre-read res_nplcm; default to NULL.}
}
\description{
This function visualizes the etiology estimates against one discrete covariate, e.g.,
age groups.
}
\seealso{
Other visualization functions: 
\code{\link{plot_BrS_panel}()},
\code{\link{plot_PERCH_regression}()},
\code{\link{plot_SS_panel}()},
\code{\link{plot_check_common_pattern}()},
\code{\link{plot_check_pairwise_SLORD}()},
\code{\link{plot_etiology_regression}()},
\code{\link{plot_etiology_side_by_side}()},
\code{\link{plot_etiology_strat}()},
\code{\link{plot_group_etiology}()},
\code{\link{plot_panels}()},
\code{\link{plot_pie_panel}()},
\code{\link{plot_selected_etiology}()},
\code{\link{plot_subwt_regression}()}
}
\concept{visualization functions}
