% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot-panels.R
\name{plot_panels}
\alias{plot_panels}
\title{Plot three-panel figures for nested partially-latent model results}
\usage{
plot_panels(DIR_NPLCM, slices = "all", bg_color = list(BrS =
  "lavenderblush", SS = "mistyrose", pie = "antiquewhite"), SS_upperlimit = 1,
  eti_upperlimit = 1, silent = TRUE)
}
\arguments{
\item{DIR_NPLCM}{File path to the folder containing posterior samples}

\item{slices}{DEFAULT is "all" - to plot all measurements; Otherwise, one can
specify a list: \code{list(MBS=c(1,3),MSS=1)} means to plot the 1st and
3rd slice of BrS measurements and 1st of SS measurement.}

\item{bg_color}{A list with names "BrS", "SS", "pie" to specify background colors<U+9286><U+FFFD>
The current default is \code{list(BrS = "lavenderblush", SS = "mistyrose",
pie="antiquewhite")}. If no background is intended, specify as NULL or for a particular
measurement, e.g., \code{BrS = NULL}.}

\item{SS_upperlimit}{The upper limit of horizontal bar for the silver-standard
subpanel (the middle panel). The default value is .25.}

\item{eti_upperlimit}{The upper limit of horizontal bar for the etiology
posterior subpanel (the rightmost panel). The default value is .4}

\item{silent}{Default is TRUE to not print any warning messages; FALSE otherwise.}
}
\value{
A figure with two or three columns
}
\description{
Visualize the model outputs for communicating how does data inform final
etiology. Current code only works for single etiologies. Multiple etiology
has not been coded, because we need special handling of pathogen ordering.
Also, because multiple etiology has a different formula for model-based observed
rate.
DN: 1. current implementation: nplcm, BrS and SS.
"Jfull" here is not the same as in other functions: it refers to the number of
pathogens even if there are pathogens with only silver-standard data;
in other functions, "Jfull" refers to the number of pathogens that have BrS data.
2. Missing data for BrS or SS are dropped when calculating observed measurement
prevalences
}

