% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot-pie-panel-two-folders.R
\name{plot_pie_panel_two_folders}
\alias{plot_pie_panel_two_folders}
\title{Plot etiology (pie) panel}
\usage{
plot_pie_panel_two_folders(model_options_list, res_nplcm_list, bugs.dat_list,
  index_for_order, dir_nplcm_list, top_pie = 1)
}
\arguments{
\item{model_options_list}{See \code{\link{nplcm}}}

\item{res_nplcm_list}{A list of posterior samples}

\item{bugs.dat_list}{Data input for the model fitting.}

\item{index_for_order}{The index of the list to be used for
ordering the pathogens. Because each model output might give
different ordering of pathogens (based on posterior mean).}

\item{dir_nplcm_list}{A list of file paths to the folders containing
posterior samples}

\item{top_pie}{Numerical value to specify the rightmost limit
on the horizontal axis for the pie panel.}
}
\description{
Now only works for singleton etiologies.
}

