% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot-SS-panel.R
\name{plot_SS_panel}
\alias{plot_SS_panel}
\title{Plot silver-standard (SS) panel}
\usage{
plot_SS_panel(slice, data_nplcm, model_options, clean_options, bugs.dat,
  res_nplcm, bg_color, top_SS = 1, cexval = 1, srtval = 0,
  prior_shape = "interval")
}
\arguments{
\item{slice}{the index of measurement slice for SS.}

\item{data_nplcm}{See \code{\link{nplcm}}}

\item{model_options}{See \code{\link{nplcm}}}

\item{clean_options}{See \code{\link{clean_perch_data}}}

\item{bugs.dat}{Data input for the model fitting.}

\item{res_nplcm}{See \code{\link{nplcm_read_folder}}}

\item{bg_color}{A list with names "BrS", "SS", "pie" to specify background colors}

\item{top_SS}{Numerical value to specify the rightmost limit
on the horizontal axis for the SS panel.}

\item{cexval}{Default is 1 - size of text of the SS percentages.}

\item{srtval}{Default is 0 - the direction of the text for the SS percentages.}

\item{prior_shape}{\code{interval} or \code{boxplot} - for how to represent
prior/posteriors of the TPR/FPRs of measurements.}
}
\description{
Plot silver-standard (SS) panel
}

