% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-etiology-regression.R
\name{plot_subwt_regression}
\alias{plot_subwt_regression}
\title{visualize the subclass weight regression with a continuous covariate}
\usage{
plot_subwt_regression(
  DIR_NPLCM,
  stratum_bool,
  case = 0,
  slice = 1,
  truth = NULL,
  RES_NPLCM = NULL
)
}
\arguments{
\item{DIR_NPLCM}{File path to the folder containing posterior samples}

\item{stratum_bool}{a vector of TRUE/FALSE with TRUE indicating the rows of subjects to include}

\item{case}{1 for plotting cases, 0 for plotting controls; default to 0.}

\item{slice}{integer; specifies which slice of bronze-standard data to visualize; Default to 1.}

\item{truth}{a list of truths computed from true parameters in simulations; elements:
Eti, FPR, PR_case,TPR; All default to \code{NULL} in real data analyses.
Currently only works for one slice of bronze-standard measurements (in a non-nested model).
\itemize{
\item truth_subwt matrix of # of rows = # of subjects, # columns: number of true subclasses
}}

\item{RES_NPLCM}{pre-read res_nplcm; default to NULL.}
}
\value{
A figure of subclass regression curves
}
\description{
visualize the subclass weight regression with a continuous covariate
}
\seealso{
Other visualization functions: 
\code{\link{plot_BrS_panel}()},
\code{\link{plot_SS_panel}()},
\code{\link{plot_check_common_pattern}()},
\code{\link{plot_check_pairwise_SLORD}()},
\code{\link{plot_etiology_regression}()},
\code{\link{plot_etiology_strat}()},
\code{\link{plot_panels}()},
\code{\link{plot_pie_panel}()}
}
\concept{visualization functions}
