% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_results.R
\name{print.nplcm}
\alias{print.nplcm}
\title{\code{print.nplcm} summarizes the results from \code{\link[=nplcm]{nplcm()}}.}
\usage{
\method{print}{nplcm}(x, ...)
}
\arguments{
\item{x}{Output from \code{\link[=nplcm]{nplcm()}}.}

\item{...}{Arguments passed to summary and printing methods.}
}
\value{
Summary of object output by \code{\link[=nplcm]{nplcm()}} --- need details.
}
\description{
\code{print.nplcm} summarizes the results from \code{\link[=nplcm]{nplcm()}}.
}
\seealso{
Other nplcm results: 
\code{\link{print.summary.nplcm.no_reg}()},
\code{\link{print.summary.nplcm.reg_nest}()},
\code{\link{print.summary.nplcm.reg_nest_strat}()},
\code{\link{print.summary.nplcm.reg_nonest}()},
\code{\link{print.summary.nplcm.reg_nonest_strat}()},
\code{\link{summary.nplcm}()}
}
\concept{nplcm results}
