% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_results.R
\name{print.summary.nplcm.reg_nonest}
\alias{print.summary.nplcm.reg_nonest}
\title{Compact printing of \code{\link[=nplcm]{nplcm()}} model fits}
\usage{
\method{print}{summary.nplcm.reg_nonest}(x, ...)
}
\arguments{
\item{x}{output from \code{summary.nplcm} with \code{summary.nplcm.reg_nonest} as the output object class.}

\item{...}{Not used.}
}
\value{
see \code{\link[=print.nplcm]{print.nplcm()}}
}
\description{
\code{print.summary.nplcm} is a print method for class
\code{summary.nplcm.reg_nonest}.
}
\seealso{
Other nplcm results: 
\code{\link{print.nplcm}()},
\code{\link{print.summary.nplcm.no_reg}()},
\code{\link{print.summary.nplcm.reg_nest}()},
\code{\link{print.summary.nplcm.reg_nest_strat}()},
\code{\link{print.summary.nplcm.reg_nonest_strat}()},
\code{\link{summary.nplcm}()}
}
\concept{nplcm results}
