% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean-perch-data.R
\name{read_meas_object}
\alias{read_meas_object}
\title{Read measurement slices}
\usage{
read_meas_object(object, data)
}
\arguments{
\item{object}{Outputs from \code{\link{make_meas_object}}}

\item{data}{Raw data with column names 
\code{\{pathogen name\}_\{specimen\}\{test\}}}
}
\value{
A list with two elements: \code{meas}-data frame with measurements;
\code{position}-see \code{\link{lookup_quality}}
}
\description{
Read measurement slices
}
\seealso{
Other raw data importing functions: 
\code{\link{extract_data_raw}()}
}
\concept{raw data importing functions}
