% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rvbern}
\alias{rvbern}
\title{Sample a vector of Bernoulli variables.}
\usage{
rvbern(p)
}
\arguments{
\item{p}{A vector of probabilities, each being the head probability
of an independent coin toss}
}
\value{
A vector of 1s (head) and 0s (tail)
}
\description{
Sample a vector of Bernoulli variables with higher speed
(same length with \code{"p"}).
The Bernoulli random variables can have different means.
}
