% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/set-prior-eti.R
\name{set_prior_eti}
\alias{set_prior_eti}
\title{Set the etiology prior}
\usage{
set_prior_eti(model_options)
}
\arguments{
\item{model_options}{See \code{nplcm_fit} function argument list.}
}
\value{
The vector of etiology prior parameters
}
\description{
Assign etiology prior to the pathogens entering the model.
}

