% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/set-prior-tpr.R
\name{set_prior_tpr_BrS}
\alias{set_prior_tpr_BrS}
\title{Set true positive rate (TPR) prior ranges for bronze-standard data.}
\usage{
set_prior_tpr_BrS(model_options, data_nplcm)
}
\arguments{
\item{model_options}{See \code{\link{nplcm}} function.}

\item{data_nplcm}{See \code{\link{assign_model}} function.}
}
\value{
Parameters for the TPR priors for BrS data. It is a list of two lists (alpha and beta).
Alpha and beta is of the same length, the number of BrS measurement slices.
Each element of the alpha (beta) list is a numeric vector for alpha (beta)
parameters to specify Beta prior for TPRs.
}
\description{
Set true positive rate (TPR) prior ranges for bronze-standard data.
}

