% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set-prior-tpr.R
\name{set_prior_tpr_SS}
\alias{set_prior_tpr_SS}
\title{Set true positive rate (TPR) prior ranges for silver-standard data.}
\usage{
set_prior_tpr_SS(model_options, data_nplcm)
}
\arguments{
\item{model_options}{See \code{\link{nplcm}} function.}

\item{data_nplcm}{See \code{\link{assign_model}} function.}
}
\value{
Parameters for the SS data TPR priors. It is a list of two lists 
(alpha and beta). Alpha and beta are of the same length, the number of 
BrS measurement slices. Each element of the alpha (beta) list is a 
numeric vector for alpha (beta) parameters to specify Beta prior for TPRs.
}
\description{
Set true positive rate (TPR) prior ranges for silver-standard data.
}
\seealso{
Other prior specification functions: \code{\link{overall_uniform}},
  \code{\link{set_prior_tpr_BrS_NoNest}}
}
