% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{show_dep}
\alias{show_dep}
\title{Show function dependencies}
\usage{
show_dep(fname, pckg = "package:baker", ...)
}
\arguments{
\item{fname}{Character string for one function}

\item{pckg}{Package name; default is \code{"package:baker"}}

\item{...}{Other parameters accepted by \code{\link[mvbutils]{foodweb}}}
}
\value{
A figure showing function dependencies
}
\description{
Show function dependencies
}
\examples{
\dontrun{
show_dep("nplcm",ancestor=FALSE)
show_dep("nplcm",ancestor=FALSE)
show_dep("nplcm_fit_NoReg",ancestor=FALSE)
show_dep("nplcm_fit_NoReg")
}

}

