% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/simulate-nplcm.R
\name{simulate_latent}
\alias{simulate_latent}
\title{Simulate Latent Status:}
\usage{
simulate_latent(set_parameter)
}
\arguments{
\item{set_parameter}{parameters for measurements}
}
\value{
a list of latent status samples for use in sampling measurements. It
also includes a template to look up measurement parameters for each type of causes.
}
\description{
Simulate Latent Status:
}

