% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/simulate-nplcm.R
\name{simulate_nplcm_known_I}
\alias{simulate_nplcm_known_I}
\title{Simulation case data from nested partially-latent class model (npLCM) family
with every individual from the same known disease class}
\usage{
simulate_nplcm_known_I(set_parameter, I_known)
}
\arguments{
\item{set_parameter}{True model parameters in the npLCM specification}

\item{I_known}{The name of cause that must be in \code{set_parameter$cause_list}}
}
\description{
Simulation case data from nested partially-latent class model (npLCM) family
with every individual from the same known disease class
}

