% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate-nplcm.R
\name{simulate_ss}
\alias{simulate_ss}
\title{Simulate Silver-Standard Data}
\usage{
simulate_ss(set_parameter, latent_samples)
}
\arguments{
\item{set_parameter}{parameters for SS measurements}

\item{latent_samples}{sampled latent status for all the subjects, for use in simulate
BrS measurements.}
}
\value{
a data frame with first column being case-control status (case at top) and
columns of silver-standard measurements
}
\description{
simulate SS measurements:
}
\seealso{
Other simulation functions: \code{\link{simulate_brs}},
  \code{\link{simulate_latent}},
  \code{\link{simulate_nplcm}}
}
