% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sort_data_frame}
\alias{sort_data_frame}
\title{Order rows of data frame according to variable combinations.}
\usage{
sort_data_frame(form, dat)
}
\arguments{
\item{form}{A Formula. See the example below;
use \code{+} for ascending, \code{-} for descending}

\item{dat}{Data frame to be ordered.}
}
\value{
Ordered data frame.
}
\description{
Order rows of data frame according to variable combinations.
}
\examples{
df <- as.data.frame(list(x1=c(2,3,1),x2=c(3,4,6)))
sort_data_frame(~-x1+x2,df)

}
