% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/eda.R
\name{summarize_BrS}
\alias{summarize_BrS}
\title{summarize bronze-standard data}
\usage{
summarize_BrS(BrS_dat, Y)
}
\arguments{
\item{BrS_dat}{bronze-stanadrd data, which is usually \code{data_nplcm$Mobs$MBS[[1]]}}

\item{Y}{A vector of case/control status: 1 for case; 0 for control}
}
\value{
a list of summaries for BrS data
}
\description{
summarize bronze-standard data
}

