% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/eda.R
\name{summarize_SS}
\alias{summarize_SS}
\title{silver-standard data summary}
\usage{
summarize_SS(SS_dat, Y)
}
\arguments{
\item{SS_dat}{a data frame of silver-standard data. It can usually
be obtained by \code{data_nplcm$Mobs$MSS[[1]]}, meaning the first SS measurement
slice.}

\item{Y}{a vector of case control status: 1 for case; 0 for control.}
}
\value{
a vector of number of positives
}
\description{
silver-standard data summary
}

