% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_results.R
\name{summary.nplcm}
\alias{summary.nplcm}
\title{\code{summary.nplcm} summarizes the results from \code{\link[=nplcm]{nplcm()}}.}
\usage{
\method{summary}{nplcm}(object, ...)
}
\arguments{
\item{object}{Output from \code{\link[=nplcm]{nplcm()}}.
An object of class "nplcm"}

\item{...}{Not used.}

\item{res_dir}{file path to folder where results are stored.}

\item{type}{\code{NoReg},\code{strat}, or \code{Reg}}
}
\value{
see \code{\link[=print.nplcm]{print.nplcm()}}
}
\description{
\code{summary.nplcm} summarizes the results from \code{\link[=nplcm]{nplcm()}}.
}
\seealso{
Other nplcm results: 
\code{\link{print.nplcm}()},
\code{\link{print.summary.nplcm.no_reg}()},
\code{\link{print.summary.nplcm.reg_nest_strat}()},
\code{\link{print.summary.nplcm.reg_nest}()},
\code{\link{print.summary.nplcm.reg_nonest_strat}()},
\code{\link{print.summary.nplcm.reg_nonest}()}
}
\concept{nplcm results}
