% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sym_diff_month}
\alias{sym_diff_month}
\title{get symmetric difference of months from two vector of R-format dates}
\usage{
sym_diff_month(Rdate1, Rdate2)
}
\arguments{
\item{Rdate1, Rdate2}{R-formated R dates. See \code{\link{as.Date}}}
}
\value{
\code{NULL} if no difference; the set of different months otherwise.
}
\description{
\code{sym_diff_month} evaluates the symmetric difference between two sets
of R-formated date
}

