% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{symb2I}
\alias{symb2I}
\title{Convert names of pathogen/combinations into 0/1 coding}
\usage{
symb2I(pathogen_name, pathogen_list)
}
\arguments{
\item{pathogen_name}{The allowed pathogen name (can be a combination of pathogens in "pathlist")}

\item{pathogen_list}{The complete list of pathogen names}
}
\value{
A 1 by length(pathlist) matrix of binary code (usually for pathogen presence/absence)
}
\description{
Convert names of pathogen/combinations into 0/1 coding
}
\examples{
symb2I("A",c("A","B","C"))
symb2I("A+B",c("A","B","C"))
symb2I("NoA",c("A","B","C"))
symb2I(c("A","B+C"),c("A","B","C")) # gives a 2 by 3 matrix.


}
