% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/TPR_prior_set.R
\name{TPR_prior_set}
\alias{TPR_prior_set}
\title{Set true positive rate (TPR) prior ranges}
\usage{
TPR_prior_set(model_options, Mobs, Y, X = NULL)
}
\arguments{
\item{model_options}{See \code{nplcm} function.}

\item{Mobs}{See \code{nplcm} function.}

\item{Y}{See \code{nplcm} function.}

\item{X}{Default is \code{NULL}. See \code{nplcm} function.}
}
\value{
Parameters for the TPR priors, separately for BrS and SS
}
\description{
Current prior assignment let bacteria NPPCR to have uniform range, viral NPPCR
to have .5-1.0 range. The PCP (a fungus) NPPCR TPR is also set to be .5-1.0; PCP
has no blood culture measurements. Also, not all the bacteria have blood culture
measurments. One question is whether to use informative NPPCR range or non-informative
NPPCR range (0-100%).
}
\details{
DN: 1.make the assignment of prior dependent on the BCX data availability
or species (bacteria)?
}

