% Generated by roxygen2 (4.0.2): do not edit by hand
\name{unfactor}
\alias{unfactor}
\title{Convert factor to numeric without losing information on the label}
\usage{
unfactor(f)
}
\arguments{
\item{f}{A factor}
}
\value{
A numeric vector
}
\description{
Convert factor to numeric without losing information on the label
}

