% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot-etiology.R
\name{unique_cause}
\alias{unique_cause}
\title{get unique causes, regardless of the actual order in combo}
\usage{
unique_cause(cause_vec)
}
\arguments{
\item{cause_vec}{a vector of characters with potential combo repetitions
written in scrambed orders separated by "+"}
}
\value{
a vector of characters with unique meanings for latent causes
}
\description{
get unique causes, regardless of the actual order in combo
}
\examples{
x <- c("A","B","A","CC+DD","DD+CC","E+F+G","B")
unique_cause(x)
}

