% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{unique_month}
\alias{unique_month}
\title{Get unique month from Date}
\usage{
unique_month(Rdate)
}
\arguments{
\item{Rdate}{standard date format in R}
}
\value{
a vector of characters with \code{month-year}, e.g., \code{4-2012}.
}
\description{
\code{unique_month} converts observed dates into unique months
to help visualize sampled months
}

