% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adhoc-imports.R
\name{write.model}
\alias{write.model}
\title{function to write bugs model (copied from R2WinBUGS)}
\usage{
write.model(model, con = "model.bug", digits = 5)
}
\arguments{
\item{model}{R / S-PLUS function containing the BUGS model in the BUGS model language, for minor differences see Section Details.}

\item{con}{passed to writeLines which actually writes the model file}

\item{digits}{number of significant digits used for WinBUGS input, see formatC}
}
\value{
write text lines to a connection; same as \code{\link[=writeLines]{writeLines()}}
}
\description{
function to write bugs model (copied from R2WinBUGS)
}
