% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/write-model-no-ppd.R
\name{write_model_NoReg_NoNest}
\alias{write_model_NoReg_NoNest}
\title{Write .bug model file for no regression, conditional independence}
\usage{
write_model_NoReg_NoNest(Mobs, prior, cause_list, use_measurements)
}
\arguments{
\item{Mobs}{measurement data in the form of \code{data_nplcm}}

\item{prior}{prior information from \code{model_options}}

\item{cause_list}{a list of latent status (crucial for building templates)}

\item{use_measurements}{"BrS", or "SS"}
}
\value{
a long character string to be written as .bug file.
}
\description{
Write .bug model file for no regression, conditional independence
}
\seealso{
\link{insert_bugfile_chunk_noreg_meas} for inserting .bug file
chunk for measurements (plug-and-play); \link{insert_bugfile_chunk_noreg_etiology}
for inserting .bug file chunk for distribution of latent status (etiology).
}

