% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/write-model-no-ppd.R
\name{write_model_NoReg_plcm}
\alias{write_model_NoReg_plcm}
\title{Write .bug model file for no regression, conditional independence}
\usage{
write_model_NoReg_plcm(Mobs, cause_list, use_measurements)
}
\arguments{
\item{Mobs}{measurements collected by `data_nplcm`}

\item{cause_list}{the vector of latent status}

\item{use_measurements}{any single or combinations of "BrS", "SS"}
}
\value{
a long character string to be written as .bug file.
}
\description{
Write .bug model file for no regression, conditional independence
}

