% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write-model.R
\name{add_meas_SS_param}
\alias{add_meas_SS_param}
\title{add parameters for a SS measurement slice among cases (conditional independence)}
\usage{
add_meas_SS_param(nslice, Mobs, prior, cause_list)
}
\arguments{
\item{nslice}{the total number of SS measurement slices}

\item{Mobs}{see \code{data_nplcm} described in \code{\link[=nplcm]{nplcm()}}}

\item{prior}{see \code{model_options} described in \code{\link[=nplcm]{nplcm()}}}

\item{cause_list}{the list of causes in \code{model_options} described in \code{\link[=nplcm]{nplcm()}}}
}
\value{
a list of two elements: the first is \code{plug}, the .bug code; the second is \code{parameters}
that stores model parameters introduced by this plugged measurement slice
}
\description{
add parameters for a SS measurement slice among cases (conditional independence)
}
\seealso{
Other likelihood specification functions: 
\code{\link{add_meas_BrS_case_Nest_Slice_jags}()},
\code{\link{add_meas_BrS_case_Nest_Slice}()},
\code{\link{add_meas_BrS_case_NoNest_Slice_jags}()},
\code{\link{add_meas_BrS_case_NoNest_Slice}()},
\code{\link{add_meas_BrS_case_NoNest_reg_Slice_jags}()},
\code{\link{add_meas_BrS_case_NoNest_reg_discrete_predictor_Slice_jags}()},
\code{\link{add_meas_BrS_ctrl_Nest_Slice}()},
\code{\link{add_meas_BrS_ctrl_NoNest_Slice}()},
\code{\link{add_meas_BrS_ctrl_NoNest_reg_Slice_jags}()},
\code{\link{add_meas_BrS_ctrl_NoNest_reg_discrete_predictor_Slice_jags}()},
\code{\link{add_meas_BrS_param_Nest_Slice_jags}()},
\code{\link{add_meas_BrS_param_Nest_Slice}()},
\code{\link{add_meas_BrS_param_Nest_reg_Slice_jags}()},
\code{\link{add_meas_BrS_param_NoNest_Slice_jags}()},
\code{\link{add_meas_BrS_param_NoNest_Slice}()},
\code{\link{add_meas_BrS_param_NoNest_reg_Slice_jags}()},
\code{\link{add_meas_BrS_param_NoNest_reg_discrete_predictor_Slice_jags}()},
\code{\link{add_meas_BrS_subclass_Nest_Slice}()},
\code{\link{add_meas_SS_case}()}

Other plug-and-play functions: 
\code{\link{add_meas_BrS_case_Nest_Slice_jags}()},
\code{\link{add_meas_BrS_case_Nest_Slice}()},
\code{\link{add_meas_BrS_case_NoNest_Slice_jags}()},
\code{\link{add_meas_BrS_case_NoNest_Slice}()},
\code{\link{add_meas_BrS_case_NoNest_reg_Slice_jags}()},
\code{\link{add_meas_BrS_case_NoNest_reg_discrete_predictor_Slice_jags}()},
\code{\link{add_meas_BrS_ctrl_Nest_Slice}()},
\code{\link{add_meas_BrS_ctrl_NoNest_Slice}()},
\code{\link{add_meas_BrS_ctrl_NoNest_reg_Slice_jags}()},
\code{\link{add_meas_BrS_ctrl_NoNest_reg_discrete_predictor_Slice_jags}()},
\code{\link{add_meas_BrS_param_Nest_Slice_jags}()},
\code{\link{add_meas_BrS_param_Nest_Slice}()},
\code{\link{add_meas_BrS_param_Nest_reg_Slice_jags}()},
\code{\link{add_meas_BrS_param_NoNest_Slice_jags}()},
\code{\link{add_meas_BrS_param_NoNest_Slice}()},
\code{\link{add_meas_BrS_param_NoNest_reg_Slice_jags}()},
\code{\link{add_meas_BrS_param_NoNest_reg_discrete_predictor_Slice_jags}()},
\code{\link{add_meas_BrS_subclass_Nest_Slice}()},
\code{\link{add_meas_SS_case}()}
}
\concept{likelihood specification functions}
\concept{plug-and-play functions}
