% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{as.matrix_or_vec}
\alias{as.matrix_or_vec}
\title{convert one column data frame to a vector}
\usage{
as.matrix_or_vec(x)
}
\arguments{
\item{x}{an one-column data.frame}
}
\value{
a vector
}
\description{
convert one column data frame to a vector
}

