% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/assign_model.R
\name{assign_model}
\alias{assign_model}
\title{Assign model based on model specification and data structure.}
\usage{
assign_model(Mobs, Y, X, model_options, silent = TRUE)
}
\arguments{
\item{Mobs}{See the \code{nplcm} function.}

\item{Y}{See the \code{nplcm} function.}

\item{X}{See the \code{nplcm} function.}

\item{model_options}{See the \code{nplcm} function.}

\item{silent}{Default is \code{TRUE}: print assigned
model descriptions on screen; otherwise, \code{FALSE}.}
}
\value{
The correct Bayesian model type to fit. The list of model
type information is:
\itemize{
\item \code{measurement}
\itemize{
\item \code{quality} e.g. "BrS+SS" indicates both BrS and SS measures are
available
\item \code{SSonly} \code{TRUE} for existence of pathogens with only SS measures;
otherwise, \code{FALSE};
\item \code{nest} \code{TRUE} for conditional dependent model; \code{FALSE}
for conditional independent model;
}
\item \code{reg}
\itemize{
\item \code{do_FPR_reg} \code{TRUE} for FPR regression; \code{FALSE} otherwise;
\item \code{do_Eti_reg} \code{TRUE} for etiology regression; \code{FALSE} otherwise;
}
}
}
\description{
Before fitting the model, this function determines a) the available measurement
types, b) false positive rate and c) etiology model components based on input
data and model options. It also does consistency checking for the specifications
in \code{model_options}, and their relations with actual inputted data.
}

