% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/assign_model.R
\name{assign_model}
\alias{assign_model}
\title{Assign model based on model specification and data.}
\usage{
assign_model(Mobs, Y, X, model_options, silent = TRUE)
}
\arguments{
\item{Mobs}{See the \code{nplcm} function.}

\item{Y}{See the \code{nplcm} function.}

\item{X}{See the \code{nplcm} function.}

\item{model_options}{See the \code{nplcm} function.}

\item{silent}{Default is \code{TRUE}: print assigned
model descriptions on screen; otherwise, \code{FALSE}.}
}
\value{
Assign the model type to be fit.
}
\description{
Determine measurements that are available and false positive rate and
etiology model components for a dataset used for fitting the model by
\link{\code{nplcm}}. It also does consistency checking for specifications
in \code{model_options}, and their relations with actual inputted data.
}

