% Generated by roxygen2 (4.0.2): do not edit by hand
\name{assign_model}
\alias{assign_model}
\title{Assign model based on model specification and data.}
\usage{
assign_model(Mobs, Y, X, model_options, silent = TRUE)
}
\arguments{
\item{Mobs}{See the \code{nplcm} function.}

\item{Y}{See the \code{nplcm} function.}

\item{X}{See the \code{nplcm} function.}

\item{model_options}{See the \code{nplcm} function.}

\item{silent}{Default is \code{TRUE}, which means print on screen assigned
model descriptions, otherwise, \code{FALSE}.}
}
\value{
Assign the model type to be fit.
}
\description{
For a data set fed into \code{\link{nplcm}}, determine measurement and
false positive rate and etiology model components. It also does consistency
checking for specifications in \code{model_options}, and their relations
with actual inputted data.
}

