% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/assign-model.R
\name{assign_model}
\alias{assign_model}
\title{Interpret the model specified by user}
\usage{
assign_model(model_options, data_nplcm, silent = TRUE)
}
\arguments{
\item{model_options}{See \code{\link{nplcm}} function.}

\item{data_nplcm}{Data for model fitting.}

\item{silent}{Default is TRUE for no messages; FALSE otherwise.}
}
\value{
A list of information for the selected model:
\itemize{
   \item \code{num_slice} a vector counting the no. of measurement slices for every
   level of measurement quality
   \item \code{nested} TRUE for nested models (conditional dependence);
   FALSE for non-nested models (conditional independence).
   \item \code{regression}
       \itemize{
           \item \code{do_reg_Eti} TRUE for doing regression on etiology (latent status);
           FALSE otherwise
           \item \code{do_reg_FPR} TRUE for doing regression on false positive rates
           (for every slice of bronze-standard); FALSE otherwise
       }
}
}
\description{
\code{assign_model} recognizes the model to fit from user input.
}
\details{
\code{assign_model} will be modified to check if data are conformable
to specified model.
}

