% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_dir_create}
\alias{check_dir_create}
\title{check existence and create folder if non-existent}
\usage{
check_dir_create(path)
}
\arguments{
\item{path}{Folder path to check and create if not there.}
}
\description{
check existence and create folder if non-existent
}
\examples{

check_dir_create(tempdir())

}
