% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/eti_prior_set.R
\name{eti_prior_set}
\alias{eti_prior_set}
\title{Set the etiology prior}
\usage{
eti_prior_set(model_options)
}
\arguments{
\item{model_options}{See \code{nplcm_fit} function argument list.}
}
\value{
The vector of etiology prior parameters
}
\description{
Assign etiology prior to the pathogens entering the model.
}

