% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/clean-perch-data.R
\name{extract_data_raw}
\alias{extract_data_raw}
\title{Convert raw data to analyzable format}
\usage{
extract_data_raw(dat_prepared, strat_nm, strat_val, meas_object,
  extra_covariates = NULL)
}
\arguments{
\item{dat_prepared}{The data set prepared in \code{clean_perch_data}}

\item{strat_nm}{The vector of covariate names to separately extract data.
For example, in PERCH data cleaning, \code{X = c("newSITE","CASECONT")}}

\item{strat_val}{The list of covariate values to stratify data.
Each element corresponds to that in \code{X}. For example, in PERCH
data cleaning, \code{Xval = list("02GAM","1")}}

\item{meas_object}{a list of bronze-standard or silver-standard measurement
objects made by function \code{\link{make_meas_object}}}

\item{extra_covariates}{The vector of covariate name for regression purposes.
  The default is NULL, which means no such covariate is necessary.}
}
\value{
A list of data.
\itemize{
\item{Mobs}{
   \itemize{
   \item{MBS} A list of bronze-standard measurements.
   The names of the list are the (specimen, test) pair. Each element of the list is a data frame.
   The rows of the data frame are subjects; the columns are (pathogen, specimen, test) trios.
   \item{MSS} A list of silver-standard measurements. Formats are the same as \code{MBS} above.
   \item{MGS} A list of gold-standard measurements. If none provided, it is \code{NULL}.
   }
}
\item{X} A data frame with columns specified by \code{extra_covariates} plus the
constant-valued \code{strat_nm} (equals \code{strat_val} by definition).
}
}
\description{
Convert raw data to analyzable format
}

