% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gateaux_download_output.R
\name{gateaux_download_output}
\alias{gateaux_download_output}
\title{Download gateaux jobs outputs from R}
\usage{
gateaux_download_output(
  report_name,
  report_id,
  JWT,
  page_no = c(0:1),
  server = "gateaux.io"
)
}
\arguments{
\item{report_name}{The name of the job}

\item{report_id}{The job ID}

\item{JWT}{String: Authentication token.}

\item{page_no}{The pages where job output are located. Defaults to the first and second pages (index 0 and 1).}

\item{server}{The gateaux server url to use. Defaults to gateaux.io}
}
\value{
JSON API return
}
\description{
Download gateaux jobs outputs from R
}
\examples{
gateaux_download_output(
  report_name = "bakeR-testreport",
  report_id = "101010",
  JWT = JWT
)
}
\author{
Dragonfly bakery
}
