% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gateaux_job_runner.R
\name{gateaux_job_runner}
\alias{gateaux_job_runner}
\title{Call arbitrary gateaux jobs from R}
\usage{
gateaux_job_runner(
  pars_list = NULL,
  report_name,
  server = "gorbachev.io",
  JWT,
  log_jobs = T,
  prefix = "jobs",
  append = T
)
}
\arguments{
\item{pars_list}{A named list of named lists of environment variables and job requirements - one list per job.
Each named list defines a job tag (name), with up to three sub-lists:
    *  pars (named environment variables passed to the job),
    *  wants (named optional job dependencies),
    *  requires (named required job dependencies).}

\item{report_name}{The name of the job}

\item{JWT}{String: Authentication token.}

\item{log_jobs}{Export job parameters to logfile?}

\item{prefix}{for the list, can include path}

\item{append}{append to existing file?}

\item{Server}{The gateaux server url to use. defaults to gorbachev.io}
}
\value{
JSON API return
}
\description{
Call arbitrary gateaux jobs from R
}
\examples{
pars = list(Run1 = list(pars = list(Env1 = "This_env", Env2 = "That_env"), wants = list(Upstream_job = 1234), requires = list(Upstream_job2 = 4321)),
            Run2 = list(pars = list(Env1 = "That_env", Env2 = "This_env"), wants = list(Upstream_job = 2222), requires = list(Upstream_job2 = 4321)))
gateaux_job_runner(pars,
                   report_name = "bakeR-testreport",
                   JWT = JWT)
}
\author{
Dragonfly bakery
}
