% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-BrS-panel.R
\name{get_fitted_mean_nested}
\alias{get_fitted_mean_nested}
\title{get fitted mean for nested model with the same subclass mixing weights}
\usage{
get_fitted_mean_nested(slice, res_nplcm, model_options, data_nplcm,
  clean_options)
}
\arguments{
\item{slice}{the slice of BrS data that are modeled}

\item{res_nplcm}{matrix of MCMC samples}

\item{model_options}{see \code{\link{nplcm}}}

\item{data_nplcm}{see \code{\link{nplcm}}}

\item{clean_options}{see \code{\link{clean_perch_data}}}
}
\value{
a matrix of no. of rows equal to retained MCMC samples, no. of columns
equal to the no. of measurement dimensions within a slice.
}
\description{
get fitted mean for nested model with the same subclass mixing weights
}
\examples{
\dontrun{
result_folder <- c("C:/2015_09_17_01KEN_nplcm")
out           <- nplcm_read_folder(result_folder)
data_nplcm    <- list(Mobs  = out$Mobs, Y = out$Y)
slice         <- 1
# fitted positive rates for pathogens separately among cases and controls:
get_fitted_mean_nested(slice,out$res_nplcm,out$model_options,data_nplcm,
                      out$clean_options)
# names of pathogens:
colnames(out$Mobs$MBS[[slice]])
}

}
