% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-metric.R
\name{get_metric}
\alias{get_metric}
\title{Obtain Integrated Squared Aitchison Distance, Squared Bias and Variance (both on 
Central Log-Ratio transformed scale) that measure the discrepancy of a posterior
distribution of pie and a true pie.}
\usage{
get_metric(DIR_NPLCM, truth)
}
\arguments{
\item{DIR_NPLCM}{File path where Bayesian results are stored}

\item{truth}{True etiologic fraction vector (must sum to 1) 
used to generate data}
}
\value{
a vector of (Integrated Squared Aitchison Distance (ISAD), bias-squared, variance, truth)
}
\description{
The result is equivalent to Euclidean-type calculation after the
compositional vector (e.g., etiologic fraction) is centered-log-ratio (CLRB) transformed.
For simulation only.
}
