% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot-etiology-side-by-side.R
\name{get_pEti_samp}
\alias{get_pEti_samp}
\title{get latent causes by names}
\usage{
get_pEti_samp(res_nplcm, model_options)
}
\arguments{
\item{res_nplcm}{result from model fits}

\item{model_options}{model specification}
}
\value{
\code{pEti_mat}: a matrix of posterior samples
}
\description{
get latent causes by names
}

