% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-panels.R
\name{get_plot_pos}
\alias{get_plot_pos}
\title{get a list of measurement index where to look for data}
\usage{
get_plot_pos(template)
}
\arguments{
\item{template}{See \code{\link{nplcm}}}
}
\value{
a list of index vectors
}
\description{
get a list of measurement index where to look for data
}
