% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getPckg}
\alias{getPckg}
\title{Get package from CRAN website}
\usage{
getPckg(pckg)
}
\arguments{
\item{pckg}{package name}
}
\value{
None
}
\description{
Get package from CRAN website
}
\examples{

\dontrun{
getPckg("ggplot2")
}
}
