% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert-bugfile-chunk.R
\name{insert_bugfile_chunk_noreg_etiology}
\alias{insert_bugfile_chunk_noreg_etiology}
\title{insert distribution for latent status code chunk into .bug file}
\usage{
insert_bugfile_chunk_noreg_etiology(ppd = NULL)
}
\arguments{
\item{ppd}{Default is NULL; set to TRUE for posterior predictive checking}
}
\value{
a long character string to be inserted into .bug model file 
as distribution specification for latent status
}
\description{
insert distribution for latent status code chunk into .bug file
}
