% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{logOR}
\alias{logOR}
\title{calculate pairwise log odds ratios}
\usage{
logOR(MBS.case, MBS.ctrl)
}
\arguments{
\item{MBS.case}{Case Bronze-Standard (BrS) data}

\item{MBS.ctrl}{Control Bronze-Standard (BrS) data}
}
\description{
Case at upper triangle; control at lower triangle
}
