% Generated by roxygen2 (4.0.2): do not edit by hand
\name{logORmat}
\alias{logORmat}
\title{Matrix of pairwise log odds ratios (LOR)}
\usage{
logORmat(MBS.case, MBS.ctrl, pathogen_display, pathogen_BrS, cex_main = 2,
  cex_se = 1)
}
\arguments{
\item{MBS.case}{Case bronze-standard measurements.}

\item{MBS.ctrl}{Control bronze-standard measurements.}

\item{pathogen_display}{The pathogen vector in desired order for display.
It can be of larger length than that of \code{pathogen_BrS}.}

\item{pathogen_BrS}{The string of pathogen names corresponding to each
row/column in the plot. All pathogens have BrS measures.}

\item{cex_main}{The size of LOR and Z-stat.}

\item{cex_se}{The size of standard error of LOR.}
}
\value{
Figure of LOR matrix and relavent s.e. and significance information.
}
\description{
The matrix of values is with cases on upper right and controls on lower
left. Log odds ratio is at the top of the cell.  Below it, its standard
error is in smaller type, using the same color as the LOR.  Then the
estimate is divided by its standard error. If it is less than 1 in
absolute value, put just a plus (red) or minus (blue) when
the Z-stat is between 1-2 in absolute value and put the actual
value when the Z is greater than 2.
}

