% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{make_template}
\alias{make_template}
\title{make a mapping template for model fitting}
\usage{
make_template(patho, cause_list)
}
\arguments{
\item{patho}{a vector of pathogen names. \code{patho}
must be substring of some \code{cause_list} elements, e.g.,
"PNEU" is a substring of "PNEU_VT13". Also see examples.}

\item{cause_list}{the list of potential latent status}
}
\value{
a mapping from \code{patho} to \code{cause_list}.
\code{NROW = length(cause_list)+1};
\code{NCOL = length(patho)}. This value is crucial in model fitting to determine
which measurements are informative of a particular category of latent status.
}
\description{
\code{make_template} creates a mapping matrix so that a measurement is mapped
to inform a particular latent status. Crucial for model fitting.
}
\details{
The first argument has to be character substrings from the second argument. The
second argument can have character strings not matched in the first argument.
}
\examples{
cause_list <- c("HINF","PNEU_VT13","PNEU_NOVT13","SAUR","HMPV_A_B","FLU_A",
"PARA_1","PARA_3","PARA_4","PV_EV","RHINO","RSV", "ENTRB","TB")

patho_BrS_NPPCR <- c("HINF","PNEU","SAUR","HMPV_A_B","FLU_A","PARA_1",
"PARA_3","PARA_4","PV_EV","RHINO","RSV")
make_template(patho_BrS_NPPCR,cause_list)


 cause = c("A","B1","B2","C","A+C","B+C")
 patho = c("A","B","C")
 make_template(patho,cause)

 cause = c("A","B1","B2","C","A+C","B+C","other")
 patho = c("A","B","C")
 make_template(patho,cause)
}

