% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/nplcm_plot_three_panel.R
\name{nplcm_plot_three_panel}
\alias{nplcm_plot_three_panel}
\title{Plot three-panel figures for nested partially-latent model results}
\usage{
nplcm_plot_three_panel(DIR_NPLCM, SS_upperlimit = 1, eti_upperlimit = 1,
  X = NULL)
}
\arguments{
\item{DIR_NPLCM}{File path to the folder containing posterior samples}

\item{SS_upperlimit}{The upper limit of horizontal bar for the silver-standard
subpanel (the middle panel). The default value is .25.}

\item{eti_upperlimit}{The upper limit of horizontal bar for the etiology
posterior subpanel (the rightmost panel). The default value is .4}

\item{X}{covariate values; default is \code{NULL}. Current use of this function
does not need X.}
}
\value{
A figure with two or three columns
}
\description{
Visualize the model outputs for communicating how does data inform final
etiology. Current code only works for single etiologies. Multiple etiology
has not been coded, because we need special handling of pathogen ordering.
Also, because multiple etiology has a different formula for model-based observed
rate.
DN: 1. current implementation: nplcm, BrS and SS.
"Jfull" here is not the same as in other functions: it refers to the number of
pathogens even if there are pathogens with only silver-standard data;
in other functions, "Jfull" refers to the number of pathogens that have BrS data.
2. Missing data for BrS or SS are dropped when calculating observed measurement
prevalences
}

