% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{parse_nplcm_reg}
\alias{parse_nplcm_reg}
\title{parse regression components (either false positive rate or etiology regression)
for fitting nplcm; Only use this when formula is non-null.}
\usage{
parse_nplcm_reg(form, data_nplcm, silent = TRUE)
}
\arguments{
\item{form}{regression formula}

\item{data_nplcm}{data object for nplcm; may contain covariates X;
must have case-control status Y.}

\item{silent}{Default is \code{TRUE} for no message about covariates;
\code{FALSE} otherwise.}
}
\value{
\code{TRUE} for doing regression; \code{FALSE} otherwise.
}
\description{
parse regression components (either false positive rate or etiology regression)
for fitting nplcm; Only use this when formula is non-null.
}
