% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{parse_nplcm_reg}
\alias{parse_nplcm_reg}
\title{parse regression components (either false positive rate or etiology regression)
for fitting nplcm}
\usage{
parse_nplcm_reg(form, data_nplcm)
}
\arguments{
\item{form}{regression formula}

\item{data_nplcm}{data object for nplcm; must contain covariates X and outcome Y.}
}
\value{
TURE for doing regression; FALSE otherwise.
}
\description{
parse regression components (either false positive rate or etiology regression)
for fitting nplcm
}

