% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/perch_data_clean.R
\name{perch_data_clean}
\alias{perch_data_clean}
\title{Clean PERCH data}
\usage{
perch_data_clean(clean_options)
}
\arguments{
\item{clean_options}{The list of options for cleaning PERCH data.
Its elements are defined as follows:
\itemize{
\item{\code{case_def}}{: variable name for case definition;}
\item{\code{case_def_val}}{: The value corresponding to cases;}
\item{\code{ctrl_def}}{: variable name for control definition;}
\item{\code{ctrl_def_val}}{: The value corresponding to controls;}
\item{\code{X_strat}}{: A vector of variable names, each defining the variables used to
stratify the data;}
\item{\code{X_strat_val}}{: A list of actual values for \code{X_strat} to
stratify the data set;}
\item{\code{pathogen_BrS_anyorder}}{: The vector of pathogen names (arbitrary order)
that have BrS measurments (For definition of Bronze-standard and
other standards, please see Wu et al. 2015, JRSS-C).
It has to be a subset of pathogen category information contained in the \code{PathCatDir}.}
\item{\code{X_extra}}{: A vector of variable names, each being the covariate name
to be included in later analysis or visualization;}
\item{\code{X_order_obs}}{: A vector of variable names, each being the covariate name
for ordering observations. For example, they can include site names or enrollment dates.
It has to be a subset of X_extra;}
\item{\code{RawMeasDir}}{: The file path to the raw data set;}
\item{\code{write_newSite}}{: Must be set to \code{TRUE} if the raw data set is changed;}
\item{\code{newSite_write_Dir}}{: The file path to a new/cleaned data set for actual
 analyses;}
\item{\code{MeasDir}}{: The file path to the cleaned data set. It is usually the same with
\code{newSite_write_Dir};}
\item{\code{PathCatDir}}{: The file path to the pathogen category information (.csv).
This list should be as complete as possible to display all pathogens used in an actual
analysis;}
\item{\code{allow_missing}}{: \code{TRUE} for using an observation that has either
BrS missing, or SS missing. Set it to \code{TRUE} if we want
to use the SS information from some cases who have missing BrS measurements. \code{TRUE}
is equivalent to using all the subjects. Missingness is judged only based on those pathogens
that are in BrS measurement assays: because if a subject miss SS measure on
a pathogen that also has BrS measure, it means SS measure is missed altogether, i.e.,
those pathogens not in BrS measurement assays cannot be detected anyway.}
}}
}
\value{
A List: \code{list(Mobs,Y,X,JSS,pathogen_MSS_ordered,pathogen_cat)}, or
with additional \code{JSSonly, pathogen_SSonly_cat} if silver-
standard only pathogens are supplied.
\itemize{
\item \code{Mobs} A list of bronze- (\code{MBS}), silver- (\code{MSS}),
and gold-standard (\code{MGS}, if available) measurements. Here if all
pathogens have BrS measures, MSS has the same number of columns as in MBS;
if some pathogens only have SS measures, then MSS will have extra columns;
\item \code{Y} 1 for case; 0 for control;
\item \code{X} Data frame of covariates for cases and controls. The names
are specified in \code{X_extra};
\item \code{JSS} Number of pathogens with both silver- and bronze-standard
data;
\item \code{pathogen_MSS_ordered} Ordered vector of pathogen names. Pathogens
with both SS and BrS pathogens are ordered first and then those with only BrS
measurements. Note that for a pathogen name vector of arbitrary order, this
function just picks out those pathogens with BrS+SS measures and puts them
at the front. Other pathogens with only BrS measures are not reordered. Pathogens with only silver-standard measures are not included.
\item \code{pathogen_cat} Pathogen categories ordered according to
\code{pathogen_MSS_ordered}.
\item \code{JSSonly} Number of pathogens with only silver-standard measures;
\item \code{pathogen_SSonly_cat} Category of pathogens with only silver-standard
data.
}
This function does not order silver-standard data only pathogens.
}
\description{
It deletes cases with two positives for BCX measures. To order cases
and controls separately according to site and enrollment date, we suggest put
\code{c("newSITE","ENRLDATE")} in \code{X_extra} and \code{X_order_obs}
whic are defined below. Currently, there must be pathogens that have both
BrS and SS measurements.
}

