% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot-pie-panel.R
\name{plot_pie_panel}
\alias{plot_pie_panel}
\title{Plot etiology (pie) panel}
\usage{
plot_pie_panel(model_options, res_nplcm, bugs.dat, bg_color, top_pie = 1,
  label_size = 1)
}
\arguments{
\item{model_options}{See \code{\link{nplcm}}}

\item{res_nplcm}{See \code{\link{nplcm_read_folder}}}

\item{bugs.dat}{Data input for the model fitting.}

\item{bg_color}{A list with names "BrS", "SS", "pie" to specify background colors}

\item{top_pie}{Numerical value to specify the rightmost limit
on the horizontal axis for the pie panel.}

\item{label_size}{the size of latent status labels on the right margin}
}
\description{
Plot etiology (pie) panel
}

