% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-group-etiology.R
\name{plot_selected_etiology}
\alias{plot_selected_etiology}
\title{Plot grouped (two groups) posterior and cascade distributions (next top-three) 
within each group}
\usage{
plot_selected_etiology(selected, DIR_NPLCM, ksFrac = 1, levellabel = 5)
}
\arguments{
\item{selected}{a vector of character strings of length 3: the names of etiologies
to visualize.}

\item{DIR_NPLCM}{The file path to the result folder}

\item{ksFrac}{A number between 0 and 1, which is the fraction of samples used to
calculate kernel stats::density}

\item{levellabel}{The contour line to be drawn in the final plot. Default is
\code{5}, which represents the contour of the \code{95} percent support region.}
}
\value{
A figure with group etiology.
}
\description{
Plot grouped (two groups) posterior and cascade distributions (next top-three) 
within each group
}
\details{
By default the function automatically orders pathogens by posterior 
means (increase clockwise)
}
\seealso{
Other visualization functions: \code{\link{plot_BrS_panel}},
  \code{\link{plot_SS_panel}},
  \code{\link{plot_check_common_pattern}},
  \code{\link{plot_check_pairwise_SLORD}},
  \code{\link{plot_etiology_regression}},
  \code{\link{plot_etiology_side_by_side}},
  \code{\link{plot_group_etiology}},
  \code{\link{plot_panels}}, \code{\link{plot_pie_panel}}
}
\concept{visualization functions}
