% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot-SS-panel.R
\name{plot_SS_panel}
\alias{plot_SS_panel}
\title{Plot silver-standard (SS) panel}
\usage{
plot_SS_panel(MSS, model_options, clean_options, res_nplcm, bugs.dat,
  top_SS = 0.3, cexval = 1, srtval = 0)
}
\arguments{
\item{MSS}{Matrix of silver-standard measurements. Rows for subjects (cases at
the top, controls at the bottom), columns for pathogen-specimen combination.
\code{MSS} has fewer columns than MBS and those pathogens with both BrS and SS
measurements should be arranged in the first several columns. This should
have been done by \code{\link{clean_perch_data}}.}

\item{model_options}{See \code{\link{nplcm}}}

\item{clean_options}{See \code{\link{clean_perch_data}}}

\item{res_nplcm}{See \code{\link{nplcm_read_folder}}}

\item{bugs.dat}{Data input for the model fitting.}

\item{top_SS}{Default is \code{0.3}. Numerical value to specify the rightmost limit
on the horizontal axis for the SS panel.}

\item{cexval}{Default is 1 - size of text of the BrS percentages.}

\item{srtval}{Default is 0 - the direction of the text for the BrS percentages.}
}
\description{
Now only works for singleton etiologies. Current the prior shape can only be
represented by intervals.
}

