% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set-prior.R
\name{set_prior_tpr_BrS_NoNest}
\alias{set_prior_tpr_BrS_NoNest}
\title{Set true positive rate (TPR) prior ranges for bronze-standard (BrS) data}
\usage{
set_prior_tpr_BrS_NoNest(slice, model_options, data_nplcm)
}
\arguments{
\item{slice}{The BrS measurement slice under consideration.}

\item{model_options}{See \code{\link[=nplcm]{nplcm()}} function.}

\item{data_nplcm}{See \code{\link[=assign_model]{assign_model()}} function.}
}
\value{
Parameters for the BrS dta TPR priors. It is a list of two lists
(alpha and beta). Alpha and beta are of the same length, the number of BrS
measurement slices. Each element of the alpha (beta) list is a numeric vector
for alpha (beta) parameters as in BETA distribution.
}
\description{
\code{set_prior_tpr_BrS_NoNest} is for for conditional independence models.
We currently also use it for conditional dependence model: subclass TPRs
are independently assigned a beta prior. Ongoing work will enable specifying
priors for the marginal TPR, i.e. TPRs for a disease class, not for the
finer subclass.
}
\seealso{
Other prior specification functions: 
\code{\link{overall_uniform}()},
\code{\link{set_prior_tpr_SS}()}
}
\concept{prior specification functions}
