% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/set-prior-tpr.R
\name{set_prior_tpr_SS}
\alias{set_prior_tpr_SS}
\title{Set true positive rate (TPR) prior ranges for silver-standard data}
\usage{
set_prior_tpr_SS(model_options, data_nplcm)
}
\arguments{
\item{model_options}{See \code{\link{nplcm}} function.}

\item{data_nplcm}{See \code{\link{assign_model}} function.}
}
\value{
Parameters for the TPR priors for SS data. It is a list of two lists (alpha and beta).
Alpha or beta is of the same length as the number of BrS measurement slices; each element of the list
is a numeric vector for alpha or beta parameters to specify Beta prior for TPRs.
}
\description{
Current prior assignment let bacteria NPPCR to have uniform range, viral NPPCR
to have .5-1.0 range. The PCP (a fungus) NPPCR TPR is also set to be .5-1.0; PCP
has no blood culture measurements. Also, not all the bacteria have blood culture
measurments. One question is whether to use informative NPPCR range or non-informative
NPPCR range (0-100%).
}

