% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/clean-perch-data.R
\name{show_individual}
\alias{show_individual}
\title{get an individual's data from the output of \code{\link{clean_perch_data}}}
\usage{
show_individual(data_nplcm, ID)
}
\arguments{
\item{data_nplcm}{data for fitting nplcm; See \link{nplcm}}

\item{ID}{patient id: \code{patid}.}
}
\value{
a list with the inquired patient's data
}
\description{
get an individual's data from the output of \code{\link{clean_perch_data}}
}

