% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eda.R
\name{show_individual}
\alias{show_individual}
\title{get an individual's data from the output of \code{\link{clean_perch_data}}}
\usage{
show_individual(data_nplcm, ID)
}
\arguments{
\item{data_nplcm}{data for fitting nplcm; See \link{nplcm}}

\item{ID}{patient id: \code{patid}.}
}
\value{
a list with the inquired patient's data
}
\description{
get an individual's data from the output of \code{\link{clean_perch_data}}
}
\examples{
\dontrun{
show_individual(data_nplcm,"G01431")
}

}
\seealso{
Other exploratory data analysis functions: \code{\link{get_top_pattern}},
  \code{\link{plot_logORmat}}, \code{\link{summarize_BrS}},
  \code{\link{summarize_SS}},
  \code{\link{visualize_season}}
}
