% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/simulate-nplcm.R
\name{simulate_nplcm}
\alias{simulate_nplcm}
\title{Simulation data from nested partially-latent class model (npLCM) family}
\usage{
simulate_nplcm(set_parameter)
}
\arguments{
\item{set_parameter}{True model parameters in the npLCM specification}
}
\value{
A list of measurements, true latent statues:
\itemize{
 \item{\code{template}} a matrix: rows for causes, columns for measurements; generated
 as a lookup table to match mixture component parmeters for every type (a particular cause) of indiviuals.
 \item{\code{datres}} all measurements along with true latent statuses
 \item{\code{dat_meas}} measurement data;
 \item{\code{dat_case}} cases' measurement data;
 \item{\code{dat_ctrl}} controls' measurement data.
 }
}
\description{
Use different case and control subclass mixing weights. Eta is of
dimension J times K. DN: document the elements in \code{set_parameter}.
}

