% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/simulate-nplcm.R
\name{simulate_nplcm}
\alias{simulate_nplcm}
\title{Simulation data from nested partially-latent class model (npLCM) family}
\usage{
simulate_nplcm(set_parameter)
}
\arguments{
\item{set_parameter}{True model parameters in the npLCM specification}
}
\description{
Use different case and control subclass mixing weights. Eta is of
dimension J times K. DN: document the elements in \code{set_parameter}.
}

