% Generated by roxygen2 (4.0.2): do not edit by hand
\name{sort_data_frame}
\alias{sort_data_frame}
\title{Order rows of data frame according to variable combinations.}
\usage{
sort_data_frame(form, dat)
}
\arguments{
\item{form}{A Formula. See example in \strong{Description}.}

\item{dat}{Data frame to be ordered.}
}
\value{
Ordered data frame.
}
\description{
Author: Kevin Wright\cr
http://tolstoy.newcastle.edu.au/R/help/04/09/4300.html
Some ideas from Andy Liaw\cr
http://tolstoy.newcastle.edu.au/R/help/04/07/1076.html\cr
Use + for ascending, - for decending.\cr
Sorting is left to right in the formula\cr
Useage is either of the following:\cr
sort.data.frame(~Block-Variety,Oats)\cr
sort.data.frame(Oats,~-Variety+Block)\cr
}

